### 2021-10-23 #3845 @whedon
[PRE REVIEW]: PySINDy: A comprehensive Python package for robust sparse system identification
**Submitting author:** @akaptano (<a href="http://orcid.org/0000-0002-6337-2907">Alan Kaptanoglu</a>)
**Repository:** <a href="https://github.com/dynamicslab/pysindy" target ="_blank">https://github.com/dynamicslab/pysindy</a>
**Version:** v1.2.3
**Reviewers:** @henrykironde, tuelwer
