### 2021-10-23 f5b806b
Okay, think I added correct DOIs
Added DOIs to the references to correct our new submission
