+++ docs/JOSS2/paper.bib
@@ -28,7 +28,8 @@
   volume={8},
   pages={169259--169271},
   year={2020},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/access.2020.3023625}
 }
 
 @article{de2020pysindy,
@@ -38,7 +39,8 @@
   volume={5},
   number={49},
   pages={1--4},
-  year={2020}
+  year={2020},
+  doi={10.21105/joss.02104}
 }
 
 @article{fasel2021sindy,
@@ -239,11 +241,16 @@
   doi = {10.1016/j.jcp.2019.07.049}
 }
 
-@article{Loiseau2019data,
-  Author = {Loiseau, Jean-Christophe},
-  Journal = {arXiv preprint arXiv:1911.07920},
-  Title = {Data-driven modeling of the chaotic thermal convection in an annular thermosyphon},
-  Year = {2019}
+@article{loiseau2019data,
+  title={Data-driven modeling of the chaotic thermal convection in an annular thermosyphon},
+  author={Loiseau, Jean-Christophe},
+  journal={Theoretical and Computational Fluid Dynamics},
+  volume={34},
+  number={4},
+  pages={339--365},
+  year={2020},
+  publisher={Springer},
+  doi = {10.1007/s00162-020-00536-w}
 }
 
 @article{Reinbold2020pre,
@@ -279,7 +286,8 @@
 	Publisher = {SIAM},
 	Title = {A survey of projection-based model reduction methods for parametric dynamical systems},
 	Volume = {57},
-	Year = {2015}
+	Year = {2015},
+	doi = {10.1137/130932715}
  }
  
  @book{Billings2013book,
@@ -301,7 +309,8 @@
 	Publisher = {National Acad Sciences},
 	Title = {Automated reverse engineering of nonlinear dynamical systems},
 	Volume = {104},
-	Year = {2007}
+	Year = {2007},
+	doi = {10.1073/pnas.0609476104}
  }
  
  @article{Brunton2021koopman,
@@ -326,7 +335,8 @@
   number={2242},
   pages={20200279},
   year={2020},
-  publisher={The Royal Society Publishing}
+  publisher={The Royal Society Publishing},
+  doi={10.1098/rspa.2020.0279}
 }
 
 @article{kaptanoglu2021promoting,
@@ -353,6 +363,7 @@
   year = {2021},
   month = {Jul},
   publisher = {American Physical Society},
+  doi = {10.1103/physreve.104.015206}
 }
 
 @book{Kutz2016book,
@@ -372,7 +383,8 @@
   number={1},
   pages={52--63},
   year={2016},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tmbmc.2016.2633265}
 }
 
 @article{messenger2021weakpde,
@@ -381,7 +393,8 @@
   journal={Journal of Computational Physics},
   pages={110525},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jcp.2021.110525}
 }
 
 @article{pathak2018model,
@@ -394,7 +407,8 @@
 	Publisher = {APS},
 	Title = {Model-free prediction of large spatiotemporally chaotic systems from data: a reservoir computing approach},
 	Volume = {120},
-	Year = {2018}
+	Year = {2018},
+	doi = {10.1103/physrevlett.120.024102}
 }
 
 @article{peherstorfer2016data,
@@ -407,6 +421,7 @@
 	Title = {Data-driven operator inference for nonintrusive projection-based model reduction},
 	Volume = {306},
 	Year = {2016},
+	doi = {10.1016/j.cma.2016.03.025}
 }
 
[…]
