### 2021-10-27 d2697dc
Add MANIFEST file to exclude example notebooks and associated data from pypi package
