+++ MANIFEST.in
@@ -1,0 +1,2 @@
+# exclude from sdist
+recursive-exclude examples *
