### 2021-10-27 ef3d6f3
Remove repeated argument test
