### 2021-10-27 15e0d88
Added basic unit test checking sample_weights works for all the optimizers.
