+++ test/optimizers/test_optimizers.py
@@ -146,6 +146,37 @@ def test_alternate_parameters(data_derivative_1d, kwargs):
     check_is_fitted(model)
 
 
+@pytest.mark.parametrize(
+    "optimizer",
+    [
+        STLSQ,
+        SSR,
+        FROLS,
+        SR3,
+        ConstrainedSR3,
+        TrappingSR3,
+    ],
+)
+def test_sample_weight_optimizers(data_lorenz, optimizer):
+    x, t = data_lorenz
+    dt = t[1] - t[0]
+    x_dot = FiniteDifference()._differentiate(x, t=dt)
+
+    sample_weight = np.ones(x[:, 0].shape)
+    sample_weight[::2] = 0
+    model = optimizer()
+    model.fit(x, x_dot)
+    model.fit(x, x_dot, sample_weight=sample_weight)
+    model.fit(x, x_dot, sample_weight=sample_weight)
+    check_is_fitted(model)
+
+    model = optimizer(normalize_columns=True)
+    model.fit(x, x_dot)
+    model.fit(x, x_dot, sample_weight=sample_weight)
+    model.fit(x, x_dot, sample_weight=sample_weight)
+    check_is_fitted(model)
+
+
 @pytest.mark.parametrize("optimizer", [STLSQ, SR3, ConstrainedSR3])
 @pytest.mark.parametrize("params", [dict(threshold=-1), dict(max_iter=0)])
 def test_general_bad_parameters(optimizer, params):
