### 2021-10-27 e82711c
Fixed the linting error and added the updated citation for Kathleens paper
