### 2021-10-30 5564329
Refactor pde_library __init__method using helper functions
