+++ pysindy/feature_library/pde_library.py
@@ -13,6 +13,7 @@ from numpy import reshape
 from numpy import shape
 from numpy import transpose
 from numpy import zeros
+from numpy import zeros_like
 from numpy.random import uniform
 from scipy.integrate import trapezoid
 from scipy.interpolate import interp1d
@@ -175,15 +176,24 @@ class PDELibrary(BaseFeatureLibrary):
         super(PDELibrary, self).__init__(
             library_ensemble=library_ensemble, ensemble_indices=ensemble_indices
         )
+        self.functions = library_functions
         self.derivative_order = derivative_order
         self.spatial_grid = spatial_grid
         self.temporal_grid = temporal_grid
-        self.functions = library_functions
         self.function_names = function_names
+        self.interaction_only = interaction_only
         self.include_bias = include_bias
         self.include_interaction = include_interaction
         self.is_uniform = is_uniform
+        self.weak_form = weak_form
+        self.K = K
         self.num_pts_per_domain = num_pts_per_domain
+        self.Hx = Hx
+        self.Hy = Hy
+        self.Hz = Hz
+        self.Ht = Ht
+        self.p = p
+
         if function_names and (len(library_functions) != len(function_names)):
             raise ValueError(
                 "library_functions and function_names must have the same"
@@ -198,288 +208,265 @@ class PDELibrary(BaseFeatureLibrary):
                 "Spatial grid and the derivative order must be "
                 "defined at the same time"
             )
-        if spatial_grid is not None and (
-            (len(spatial_grid.shape) != 1)
-            and (len(spatial_grid.shape) != 3)
-            and (len(spatial_grid.shape) != 4)
-        ):
+        if spatial_grid is not None and (len(spatial_grid.shape) not in (1, 3, 4)):
             raise ValueError("Spatial grid size is incorrect")
-        self.interaction_only = interaction_only
         if spatial_grid is not None:
-            self.s_len = len((self.spatial_grid).shape)
-            if self.s_len == 1:
+            self.space_ndim = len((self.spatial_grid).shape)
+            if self.space_ndim == 1:
                 self.num_derivatives = derivative_order
-            elif self.s_len == 3:
+            elif self.space_ndim == 3:
                 num_derivatives = 2
                 for i in range(2, derivative_order + 1):
                     num_derivatives += i + 1
                 self.num_derivatives = num_derivatives
-            elif self.s_len == 4:
+            elif self.space_ndim == 4:
                 k = 3
                 num_derivatives = 0
                 for i in range(1, derivative_order + 1):
                     num_derivatives += int(n_choose_k(i + k - 1, k - 1))
                 self.num_derivatives = num_derivatives
         else:
-            self.s_len = 0
+            self.space_ndim = 0
 
-        # weak form checks now
-        if weak_form and temporal_grid is not None:
-            if len(shape(temporal_grid)) != 1:
+        # Weak form checks and setup
+        if self.weak_form:
+            self._weak_form_setup()
+
+    def _weak_form_setup(self):
+        # Assumes self.weak_form = True
+        if self.temporal_grid is None:
+            raise ValueError("self.temporal_grid cannot be None if weak_form=True")
+        else:
+            if len(shape(self.temporal_grid)) != 1:
                 raise ValueError("Temporal grid must be 1D.")
[…]
