### 2021-10-30 6231489
Update criteria docstring entry
