+++ pysindy/pysindy.py
@@ -5,6 +5,7 @@ from numpy import concatenate
 from numpy import copy
 from numpy import insert
 from numpy import isscalar
+from numpy import median
 from numpy import ndim
 from numpy import newaxis
 from numpy import sort
@@ -190,6 +191,7 @@ class SINDy(BaseEstimator):
         n_candidates_to_drop=1,
         n_subset=None,
         n_models=None,
+        ensemble_aggregator=None,
     ):
         """
         Fit a SINDy model.
@@ -270,6 +272,13 @@ class SINDy(BaseEstimator):
         n_models : int, optional (default 20)
             Number of models to generate via ensemble
 
+        ensemble_aggregator : callable, optional (default numpy.median)
+            Method to aggregate model coefficients across different samples.
+            This method argument is only used if ``ensemble`` or ``library_ensemble``
+            is True.
+            The method should take in a list of 2D arrays and return a 2D
+            array of the same shape as the arrays in the list.
+
         Returns
         -------
         self: a fitted :class:`SINDy` instance
@@ -324,7 +333,7 @@ class SINDy(BaseEstimator):
                 else:
                     x_dot = validate_input(x_dot, t)
 
-        # set ensemble variables
+        # Set ensemble variables
         self.ensemble = ensemble
         self.library_ensemble = library_ensemble
 
@@ -364,18 +373,22 @@ class SINDy(BaseEstimator):
                         replace,
                         tgrid,
                         self.feature_library,
-                        PDELibrary,
+                        isinstance(self.feature_library, PDELibrary),
                     )
                     self.model.fit(x_ensemble, x_dot_ensemble)
                     self.coef_list.append(self.model.steps[-1][1].coef_)
             elif library_ensemble and not ensemble:
                 self.feature_library.library_ensemble = True
                 (self.feature_library).fit(x)
-                N = self.feature_library.n_output_features_
+                n_output_features = self.feature_library.n_output_features_
                 self.coef_list = []
                 for i in range(n_models):
                     self.feature_library.ensemble_indices = sort(
-                        choice(range(N), n_candidates_to_drop, replace=False)
+                        choice(
+                            range(n_output_features),
+                            n_candidates_to_drop,
+                            replace=False,
+                        )
                     )
                     self.model.fit(x, x_dot)
                     coef_partial = self.model.steps[-1][1].coef_
@@ -390,7 +403,7 @@ class SINDy(BaseEstimator):
             elif ensemble and library_ensemble:
                 self.feature_library.library_ensemble = True
                 (self.feature_library).fit(x)
-                N = self.feature_library.n_output_features_
+                n_output_features = self.feature_library.n_output_features_
                 self.coef_list = []
                 for i in range(n_models):
                     x_ensemble, x_dot_ensemble = drop_random_rows(
@@ -400,11 +413,15 @@ class SINDy(BaseEstimator):
                         replace,
                         tgrid,
                         self.feature_library,
-                        PDELibrary,
+                        isinstance(self.feature_library, PDELibrary),
                     )
                     for j in range(n_models):
                         self.feature_library.ensemble_indices = sort(
-                            choice(range(N), n_candidates_to_drop, replace=False)
+                            choice(
+                                range(n_output_features),
+                                n_candidates_to_drop,
+                                replace=False,
+                            )
                         )
                         self.model.fit(x_ensemble, x_dot_ensemble)
                         coef_partial = self.model.steps[-1][1].coef_
@@ -419,7 +436,14 @@ class SINDy(BaseEstimator):
             else:
                 self.model.fit(x, x_dot)
 
-        # Annoying sklearn change that have to replace these everywhere
+        # Get average coefficients if ensembling was used
[…]
