### 2021-11-06 cbf27d3
Added a new library for doing spatiotemporal functions. Made most of Brians suggested changes.
