+++ pysindy/feature_library/spatiotemporal_library.py
@@ -1,0 +1,305 @@
+from itertools import combinations
+from itertools import combinations_with_replacement as combinations_w_r
+
+from numpy import empty
+from numpy import ones
+from numpy import reshape
+from numpy import shape
+from numpy import tile
+from numpy import zeros
+from sklearn import __version__
+from sklearn.utils import check_array
+from sklearn.utils.validation import check_is_fitted
+
+from .base import BaseFeatureLibrary
+
+
+class SpatiotemporalLibrary(BaseFeatureLibrary):
+    """
+    Generate a custom library from space and time variables.
+    Allows for the spatial and temporal variables to all have
+    differing dimensions for non-uniform, non-square meshes.
+    Note that this is quite a different library than others,
+    which generate candidate library terms from
+    the input data, not from the space-time variables!
+
+    Parameters
+    ----------
+    spatiotemporal_variables : list, shape (n_spatial_dims + 1)
+        The spatiotemporal grid for building library terms that are functions
+        of the spatiotemporal variables (x, y, z, ..., t).
+        The grid can be non-uniform and non-square. For 3D spatial grid,
+        this might look like spatiotemporal_variables = [x, y, z, t]
+        (although the order of the terms does not matter), where x is a
+        numpy array of dimension (nx, ny, nz, nt), (same for y, z, t)
+        Note that spatiotemporal_variables is not a numpy array since that
+        would require x, y, z, ..., t to all have the same dimensions.
+
+    library_functions : list of mathematical functions
+        Functions to include in the library. Default is to use same functions
+        for all variables. Can also be used so that each variable has an
+        associated library, in this case library_functions is shape
+        (n_input_features, num_library_functions)
+
+    function_names : list of functions, optional (default None)
+        List of functions used to generate feature names for each library
+        function. Each name function must take a string input (representing
+        a variable name), and output a string depiction of the respective
+        mathematical function applied to that variable. For example, if the
+        first library function is sine, the name function might return
+        :math:`\\sin(x)` given :math:`x` as input. The function_names list
+        must be the same length as library_functions. If no list of function
+        names is provided, defaults to using
+        :math:`[ f_0(x),f_1(x), f_2(x), \\ldots ]`.
+
+    interaction_only : boolean, optional (default True)
+        Whether to omit self-interaction terms.
+        If True, function evaulations of the form :math:`f(x,x)` and :math:`f(x,y,x)`
+        will be omitted, but those of the form :math:`f(x,y)` and :math:`f(x,y,z)`
+        will be included.
+        If False, all combinations will be included.
+
+    library_ensemble : boolean, optional (default False)
+        Whether or not to use library bagging (regress on subset of the
+        candidate terms in the library)
+
+    ensemble_indices : integer array, optional (default [0])
+        The indices to use for ensembling the library.
+
+    include_bias : boolean, optional (default False)
+        If True (default), then include a bias column, the feature in which
+        all polynomial powers are zero (i.e. a column of ones - acts as an
+        intercept term in a linear model).
+        This is hard to do with just lambda functions, because if the system
+        is not 1D, lambdas will generate duplicates.
+
+    Attributes
+    ----------
+    functions : list of functions
+        Mathematical library functions to be applied to each input feature.
+
+    function_names : list of functions
+        Functions for generating string representations of each library
+        function.
+
+    n_input_features_ : int
+        The total number of input features.
+        WARNING: This is deprecated in scikit-learn version 1.0 and higher so
+        we check the sklearn.__version__ and switch to n_features_in if needed.
+
+    n_output_features_ : int
[…]
