### 2021-11-06 70d55ec
Make spatiovariable check more robust
