### 2021-11-06 e4a2683
Update paper reference in docstring
