### 2021-11-06 fd7ac54
Improve docstring for ensemble_aggregator
