### 2021-11-06 81a567a
Refactor bad parameter tests
