+++ test/feature_library/test_feature_library.py
@@ -118,212 +118,109 @@ def test_bad_parameters():
         CustomLibrary(
             library_functions=library_functions, function_names=function_names
         )
+
+
+def test_spatiotemporal_library_incorrect_grid_size():
+    library_functions = [lambda x: x, lambda x: x ** 2, lambda x: 0 * x]
+    function_names = [lambda s: str(s), lambda s: "{}^2".format(s)]
+    x = np.linspace(0, 10, 10)
+    y = np.linspace(0, 10, 20)
+    X, Y = np.meshgrid(x, y, indexing="ij")
+
     with pytest.raises(ValueError):
-        library_functions = [lambda x: x, lambda x: x ** 2, lambda x: 0 * x]
-        function_names = [lambda s: str(s), lambda s: "{}^2".format(s)]
-        x = np.linspace(0, 10, 10)
-        y = np.linspace(0, 10, 20)
-        X, Y = np.meshgrid(x, y, indexing="ij")
         SpatiotemporalLibrary(
             library_functions=library_functions,
             function_names=function_names,
             spatiotemporal_variables=[X, Y],
         )
+
     with pytest.raises(ValueError):
-        library_functions = [lambda x: x, lambda x: x ** 2]
         SpatiotemporalLibrary(
             library_functions=library_functions,
-            spatiotemporal_variables=np.linspace(0, 10, 10),
+            spatiotemporal_variables=x,
         )
+
     with pytest.raises(ValueError):
-        library_functions = [lambda x: x, lambda x: x ** 2]
-        X = np.linspace(0, 10, 10)
-        Y = np.linspace(0, 10, 10)
         SpatiotemporalLibrary(
-            library_functions=library_functions, spatiotemporal_variables=[X, Y]
+            library_functions=library_functions, spatiotemporal_variables=[x, y]
         )
-    with pytest.raises(ValueError):
-        library_functions = [lambda x: x, lambda x: x ** 2, lambda x: 0 * x]
-        function_names = [lambda s: str(s), lambda s: "{}^2".format(s)]
-        PDELibrary(library_functions=library_functions, function_names=function_names)
-    with pytest.raises(ValueError):
-        library_functions = [lambda x: x, lambda x: x ** 2, lambda x: 0 * x]
-        PDELibrary(library_functions=library_functions, derivative_order=-1)
-    with pytest.raises(ValueError):
-        library_functions = [lambda x: x, lambda x: x ** 2, lambda x: 0 * x]
-        PDELibrary(library_functions=library_functions, spatial_grid=range(10))
-    with pytest.raises(ValueError):
-        library_functions = [lambda x: x, lambda x: x ** 2, lambda x: 0 * x]
-        PDELibrary(library_functions=library_functions, derivative_order=3)
-    with pytest.raises(ValueError):
-        library_functions = [lambda x: x, lambda x: x ** 2, lambda x: 0 * x]
-        PDELibrary(
-            library_functions=library_functions,
-            spatial_grid=range(10),
-            derivative_order=-1,
-        )
-    with pytest.raises(ValueError):
-        library_functions = [lambda x: x, lambda x: x ** 2, lambda x: 0 * x]
-        PDELibrary(
-            library_functions=library_functions,
-            spatial_grid=np.zeros((10, 10)),
-        )
-    with pytest.raises(ValueError):
-        library_functions = [lambda x: x, lambda x: x ** 2, lambda x: 0 * x]
-        PDELibrary(
-            library_functions=library_functions,
-            spatial_grid=np.zeros((10, 10, 10, 10, 10)),
-        )
-    with pytest.raises(ValueError):
[…]
