### 2021-11-07 2185a8f
Refactor pde and spatiotemporal library tests
