### 2021-11-13 84830d0
Added a library for tensoring two libraries together. Added a GeneralizedLibrary that can basically make any library one wants, with any subset of the inputs one wants. Removed SpatiotemporalLibrary.
