+++ pysindy/feature_library/__init__.py
@@ -1,19 +1,21 @@
 from .base import ConcatLibrary
+from .base import GeneralizedLibrary
+from .base import TensoredLibrary
 from .custom_library import CustomLibrary
 from .fourier_library import FourierLibrary
 from .identity_library import IdentityLibrary
 from .pde_library import PDELibrary
 from .polynomial_library import PolynomialLibrary
 from .sindy_pi_library import SINDyPILibrary
-from .spatiotemporal_library import SpatiotemporalLibrary
 
 __all__ = [
     "ConcatLibrary",
+    "TensoredLibrary",
+    "GeneralizedLibrary",
     "CustomLibrary",
     "FourierLibrary",
     "IdentityLibrary",
     "PolynomialLibrary",
     "PDELibrary",
     "SINDyPILibrary",
-    "SpatiotemporalLibrary",
 ]
