+++ pysindy/feature_library/base.py
@@ -106,6 +106,12 @@ class BaseFeatureLibrary(TransformerMixin):
     def __add__(self, other):
         return ConcatLibrary([self, other])
 
+    def __mul__(self, other):
+        return TensoredLibrary([self, other])
+
+    def __rmul__(self, other):
+        return TensoredLibrary([self, other])
+
     @property
     def size(self):
         check_is_fitted(self)
@@ -253,6 +259,503 @@ class ConcatLibrary(BaseFeatureLibrary):
         """
         feature_names = list()
         for lib in self.libraries_:
+            lib_feat_names = lib.get_feature_names(input_features)
+            feature_names += lib_feat_names
+        return feature_names
+
+
+class TensoredLibrary(BaseFeatureLibrary):
+    """Tensor multiple libraries together into one library. All settings
+    provided to individual libraries will be applied.
+
+    Parameters
+    ----------
+    libraries : list of libraries
+        Library instances to be applied to the input matrix.
+
+    library_ensemble : boolean, optional (default False)
+        Whether or not to use library bagging (regress on subset of the
+        candidate terms in the library).
+
+    ensemble_indices : integer array, optional (default [0])
+        The indices to use for ensembling the library. For instance, if
+        ensemble_indices = [0], it chops off the first column of the library.
+
+    Attributes
+    ----------
+    libraries_ : list of libraries
+        Library instances to be applied to the input matrix.
+
+    inputs_per_library_ : numpy nd.array (default None)
+        Array that specifies which inputs should be used for each of the
+        libraries you are going to tensor together
+
+    n_input_features_ : int
+        The total number of input features.
+        WARNING: This is deprecated in scikit-learn version 1.0 and higher so
+        we check the sklearn.__version__ and switch to n_features_in if needed.
+
+    n_output_features_ : int
+        The total number of output features. The number of output features
+        is the product of the numbers of output features for each of the
+        libraries that were tensored together.
+
+    Examples
+    --------
+    >>> import numpy as np
+    >>> from pysindy.feature_library import FourierLibrary, CustomLibrary
+    >>> from pysindy.feature_library import TensoredLibrary
+    >>> x = np.array([[0.,-1],[1.,0.],[2.,-1.]])
+    >>> functions = [lambda x : np.exp(x), lambda x,y : np.sin(x+y)]
+    >>> lib_custom = CustomLibrary(library_functions=functions)
+    >>> lib_fourier = FourierLibrary()
+    >>> lib_tensored = TensoredLibrary([lib_custom, lib_fourier])
+    >>> lib_tensored.fit()
+    >>> lib.transform(x)
+    """
+
+    def __init__(
+        self,
+        libraries: list,
+        library_ensemble=False,
+        inputs_per_library=None,
+        ensemble_indices=[0],
+    ):
+        super(TensoredLibrary, self).__init__(
+            library_ensemble=library_ensemble, ensemble_indices=ensemble_indices
+        )
+        self.libraries_ = libraries
+        self.inputs_per_library_ = inputs_per_library
+
+    def _combinations(self, lib_i, lib_j):
+        """
+        Compute combinations of the numerical libraries.
+
+        Returns
+        -------
+        lib_full : All combinations of the numerical library terms.
+        """
+        lib_full = np.zeros((lib_i.shape[0], lib_i.shape[-1] * lib_j.shape[-1]))
+        q = 0
+        for i in range(lib_i.shape[-1]):
+            for j in range(lib_j.shape[-1]):
+                lib_full[:, q] = lib_i[:, i] * lib_j[:, j]
+                q = q + 1
+        return lib_full
+
+    def _name_combinations(self, lib_i, lib_j):
+        """
[…]
