### 2021-11-13 4e797cf
Added tiny test of the TensoredLibrary class
