### 2021-11-14 1c3272c
Small changes to the documentation of new library.
