+++ pysindy/feature_library/base.py
@@ -288,7 +288,8 @@ class TensoredLibrary(BaseFeatureLibrary):
 
     inputs_per_library_ : numpy nd.array (default None)
         Array that specifies which inputs should be used for each of the
-        libraries you are going to tensor together
+        libraries you are going to tensor together. Used for building
+        GeneralizedLibrary objects.
 
     n_input_features_ : int
         The total number of input features.
@@ -309,9 +310,9 @@ class TensoredLibrary(BaseFeatureLibrary):
     >>> functions = [lambda x : np.exp(x), lambda x,y : np.sin(x+y)]
     >>> lib_custom = CustomLibrary(library_functions=functions)
     >>> lib_fourier = FourierLibrary()
-    >>> lib_tensored = TensoredLibrary([lib_custom, lib_fourier])
-    >>> lib_tensored.fit()
-    >>> lib.transform(x)
+    >>> lib_tensored = lib_custom * lib_fourier
+    >>> lib_tensored.fit(x)
+    >>> lib_tensored.transform(x)
     """
 
     def __init__(
@@ -359,7 +360,7 @@ class TensoredLibrary(BaseFeatureLibrary):
 
     def _set_inputs_per_library(self, inputs_per_library):
         """
-        Extra function to make GeneralizedLibrary work easier
+        Extra function to make building a GeneralizedLibrary object easier
         """
         self.inputs_per_library_ = inputs_per_library
 
@@ -494,7 +495,9 @@ class GeneralizedLibrary(BaseFeatureLibrary):
     """Put multiple libraries into one library. All settings
     provided to individual libraries will be applied. Note that this class
     allows one to specifically choose which input variables are used for
-    each library, and take tensor products of any pair of libraries.
+    each library, and take tensor products of any pair of libraries. Tensored
+    libraries inherit the same input variables specified for the individual
+    libraries.
 
     Parameters
     ----------
@@ -533,6 +536,25 @@ class GeneralizedLibrary(BaseFeatureLibrary):
     libraries_ : list of libraries
         Library instances to be applied to the input matrix.
 
+    tensor_array_ : 2D list of booleans, optional,
+            (default is to not tensor any of the libraries together) shape
+            equal to (# of tensor libraries to make, # feature libraries)
+        Indicates which pairs of libraries to tensor product together and
+        add to the overall library. For instance if you have 5 libraries,
+        and want to do two tensor products, you could use the list
+        [[1, 0, 0, 1, 0], [0, 1, 0, 1, 1]] to indicate that you want two
+        tensored libraries from tensoring libraries 0 and 3 and libraries
+        1, 3, and 4.
+
+    inputs_per_library_ : np.ndarray, optional
+            (default all inputs used for every library) shape equal to
+            (# feature libraries, # variable inputs)
+        Can be used to specify a subset of the variables to use to generate
+        a feature library. If number of feature libraries > 1, then can be
+        use to generate a large number of libraries, each using their own
+        subsets of the input variables. Note that this must be specified for
+        all the individual feature libraries.
+
     n_input_features_ : int
         The total number of input features.
         WARNING: This is deprecated in scikit-learn version 1.0 and higher so
@@ -553,8 +575,8 @@ class GeneralizedLibrary(BaseFeatureLibrary):
     >>> lib_custom = CustomLibrary(library_functions=functions)
     >>> lib_fourier = FourierLibrary()
     >>> lib_generalized = GeneralizedLibrary([lib_custom, lib_fourier])
-    >>> lib_generalized.fit()
-    >>> lib.transform(x)
+    >>> lib_generalized.fit(x)
+    >>> lib_generalized.transform(x)
     """
 
     def __init__(
[…]
