### 2021-11-17 3c39cc4
Update README
Added documentation about new paper on arxiv
