### 2021-11-17 582f48e
Update paper.bib
Added Charles's new paper to biblio
