+++ docs/JOSS2/paper.bib
@@ -1,3 +1,10 @@
+@article{delahunt2021toolkit,
+  title={A toolkit for data-driven discovery of governing equations in high-noise regimes},
+  author={Delahunt, Charles B and Kutz, J Nathan},
+  journal={arXiv preprint arXiv:2111.04870},
+  year={2021}
+}
+
 @article{brunton2016pnas,
   title = {Discovering governing equations from data by sparse identification of nonlinear dynamical systems},
   author = {Brunton, Steven L. and Proctor, Joshua L. and Kutz, J. Nathan},
