### 2021-11-17 b217013
Update paper.md
Added Charles's new paper reference to new JOSS paper
