### 2021-11-19 0b9d6ce
Fix typo in docstring
