### 2021-11-19 9ebeda9
Remove copy-paste artifacts from attribute docstrings
