### 2021-11-19 ec102eb
Move GeneralizedLibrary class to its own file
