+++ pysindy/feature_library/__init__.py
@@ -1,8 +1,8 @@
 from .base import ConcatLibrary
-from .base import GeneralizedLibrary
 from .base import TensoredLibrary
 from .custom_library import CustomLibrary
 from .fourier_library import FourierLibrary
+from .generalized_library import GeneralizedLibrary
 from .identity_library import IdentityLibrary
 from .pde_library import PDELibrary
 from .polynomial_library import PolynomialLibrary
