### 2021-11-19 87b13e5
Refactor TrappingSR3 quadratic library test
