### 2021-11-19 c340065
Move GeneralizedLibrary to its own file
