+++ pysindy/feature_library/generalized_library.py
@@ -1,0 +1,298 @@
+import numpy as np
+from sklearn import __version__
+from sklearn.utils import check_array
+from sklearn.utils.validation import check_is_fitted
+
+from .base import BaseFeatureLibrary
+
+
+class GeneralizedLibrary(BaseFeatureLibrary):
+    """Put multiple libraries into one library. All settings
+    provided to individual libraries will be applied. Note that this class
+    allows one to specifically choose which input variables are used for
+    each library, and take tensor products of any pair of libraries. Tensored
+    libraries inherit the same input variables specified for the individual
+    libraries.
+
+    Parameters
+    ----------
+    libraries : list of libraries
+        Library instances to be applied to the input matrix.
+
+    tensor_array : 2D list of booleans, optional,
+            (default is to not tensor any of the libraries together) shape
+            equal to (# of tensor libraries to make, # feature libraries)
+        Indicates which pairs of libraries to tensor product together and
+        add to the overall library. For instance if you have 5 libraries,
+        and want to do two tensor products, you could use the list
+        [[1, 0, 0, 1, 0], [0, 1, 0, 1, 1]] to indicate that you want two
+        tensored libraries from tensoring libraries 0 and 3 and libraries
+        1, 3, and 4.
+
+    inputs_per_library : np.ndarray, optional
+            (default all inputs used for every library) shape equal to
+            (# feature libraries, # variable inputs)
+        Can be used to specify a subset of the variables to use to generate
+        a feature library. If number of feature libraries > 1, then can be
+        used to generate a large number of libraries, each using their own
+        subsets of the input variables. Note that this must be specified for
+        all the individual feature libraries.
+
+    library_ensemble : boolean, optional (default False)
+        Whether or not to use library bagging (regress on subset of the
+        candidate terms in the library).
+
+    ensemble_indices : integer array, optional (default [0])
+        The indices to use for ensembling the library. For instance, if
+        ensemble_indices = [0], it chops off the first column of the library.
+
+    Attributes
+    ----------
+    libraries_ : list of libraries
+        Library instances to be applied to the input matrix.
+
+    tensor_array_ : 2D list of booleans
+        Indicates which pairs of libraries to tensor product together and
+        add to the overall library. For instance if you have 5 libraries,
+        and want to do two tensor products, you could use the list
+        [[1, 0, 0, 1, 0], [0, 1, 0, 1, 1]] to indicate that you want two
+        tensored libraries from tensoring libraries 0 and 3 and libraries
+        1, 3, and 4. Shape equal to
+        (# of tensor libraries to make, # feature libraries)
+
+    inputs_per_library_ : np.ndarray
+        Can be used to specify a subset of the variables to use to generate
+        a feature library. If number of feature libraries > 1, then can be
+        use to generate a large number of libraries, each using their own
+        subsets of the input variables. Note that this must be specified for
+        all the individual feature libraries. Shape equal to
+            (# feature libraries, # variable inputs)
+
+    n_input_features_ : int
+        The total number of input features.
+        WARNING: This is deprecated in scikit-learn version 1.0 and higher so
+        we check the sklearn.__version__ and switch to n_features_in if needed.
+
+    n_output_features_ : int
+        The total number of output features. The number of output features
+        is the sum of the numbers of output features for each of the
+        concatenated libraries.
+
+    Examples
+    --------
+    >>> import numpy as np
+    >>> from pysindy.feature_library import FourierLibrary, CustomLibrary
+    >>> from pysindy.feature_library import GeneralizedLibrary
+    >>> x = np.array([[0.,-1],[1.,0.],[2.,-1.]])
+    >>> functions = [lambda x : np.exp(x), lambda x,y : np.sin(x+y)]
[…]
