### 2021-11-19 c22c079
Removed some older code that was defunct. Reran 1_feature_overview to make sure still works with changes to generalizedlibrary.
