### 2021-11-19 a4d454b
Fixed a bunch of small errors, including incorrect prox for weighted L1 norm, old code in the trapping optimizer, some failing tests, and so on. Added tests of the constrained optimizers that actually tests if the constraints are correct. Got rid of the torrent of cvxpy warnings during pytest.
