### 2021-11-19 ade5bdf
Pytest not passing online. Tried to fix this with some small change in test_optimizers.
