### 2021-11-20 2e1a61f
Added flow chart and sentence to the second JOSS paper
