### 2021-11-22 b8e723b
Made a number of plotting and documentation and other changes to the example notebooks as per Brians recent comments on the pull request. Also changed basically everything to solve_ivp (with a few exceptions because odeint only sends warnings about unstable models while solve_ivp crashes and burns), which turned out to be a lot of changes in the notebooks and the tests. Made some small edits to the SSR algorithm so it continues until no more coefficients can be chopped. Added the logistic map system to the odes.py file.
