+++ test/feature_library/test_feature_library.py
@@ -785,7 +785,7 @@ def test_sindypi_library(data_lorenz):
     model.fit(x, t=t)
     assert np.shape(sindy_opt.coef_) == (40, 40)
 
-    sindy_opt = SINDyPI(threshold=0.1, thresholder="l1", model_subset=[3])
+    sindy_opt = SINDyPI(threshold=1, thresholder="l1", model_subset=[3])
     model = SINDy(
         optimizer=sindy_opt,
         feature_library=sindy_library,
