+++ pysindy/utils/odes.py
@@ -2,12 +2,12 @@ import numpy as np
 
 
 # Linear, damped harmonic oscillator
-def linear_damped_SHO(x, t):
+def linear_damped_SHO(t, x):
     return [-0.1 * x[0] + 2 * x[1], -2 * x[0] - 0.1 * x[1]]
 
 
 # Cubic, damped harmonic oscillator
-def cubic_damped_SHO(x, t):
+def cubic_damped_SHO(t, x):
     return [
         -0.1 * x[0] ** 3 + 2 * x[1] ** 3,
         -2 * x[0] ** 3 - 0.1 * x[1] ** 3,
@@ -15,45 +15,60 @@ def cubic_damped_SHO(x, t):
 
 
 # Linear 3D toy system
-def linear_3D(x, t):
+def linear_3D(t, x):
     return [-0.1 * x[0] + 2 * x[1], -2 * x[0] - 0.1 * x[1], -0.3 * x[2]]
 
 
 # Van der Pol ODE
-def van_der_pol(x, t, p=[0.5]):
+def van_der_pol(t, x, p=[0.5]):
     return [x[1], p[0] * (1 - x[0] ** 2) * x[1] - x[0]]
 
 
 # Duffing ODE
-def duffing(x, t, p=[0.2, 0.05, 1]):
+def duffing(t, x, p=[0.2, 0.05, 1]):
     return [x[1], -p[0] * x[1] - p[1] * x[0] - p[2] * x[0] ** 3]
 
 
 # Lotka model
-def lotka(x, t, p=[1, 10]):
+def lotka(t, x, p=[1, 10]):
     return [p[0] * x[0] - p[1] * x[0] * x[1], p[1] * x[0] * x[1] - 2 * p[0] * x[1]]
 
 
 # Generic cubic oscillator model
-def cubic_oscillator(x, t, p=[-0.1, 2, -2, -0.1]):
+def cubic_oscillator(t, x, p=[-0.1, 2, -2, -0.1]):
     return [p[0] * x[0] ** 3 + p[1] * x[1] ** 3, p[2] * x[0] ** 3 + p[3] * x[1] ** 3]
 
 
 # Rossler model
-def rossler(x, t, p=[0.2, 0.2, 5.7]):
+def rossler(t, x, p=[0.2, 0.2, 5.7]):
     return [-x[1] - x[2], x[0] + p[0] * x[1], p[1] + (x[0] - p[2]) * x[2]]
 
 
 # Hopf bifurcation model
-def hopf(x, t, mu=-0.05, omega=1, A=1):
+def hopf(t, x, mu=-0.05, omega=1, A=1):
     return [
         mu * x[0] - omega * x[1] - A * x[0] * (x[0] ** 2 + x[1] ** 2),
         omega * x[0] + mu * x[1] - A * x[1] * (x[0] ** 2 + x[1] ** 2),
     ]
 
 
+# Logistic map model
+def logistic_map(x, mu):
+    return mu * x * (1 - x)
+
+
+# Logistic map model with linear control input
+def logistic_map_control(x, mu, u):
+    return mu * x * (1 - x) + u
+
+
+# Logistic map model with other control input
+def logistic_map_multicontrol(x, mu, u):
+    return mu * x * (1 - x) + u[0] * u[1]
+
+
 # Lorenz model
-def lorenz(x, t, sigma=10, beta=2.66667, rho=28):
+def lorenz(t, x, sigma=10, beta=2.66667, rho=28):
     return [
         sigma * (x[1] - x[0]),
         x[0] * (rho - x[2]) - x[1],
@@ -67,7 +82,8 @@ def lorenz_u(t):
 
 
 # Lorenz equations with control input
[…]
