### 2021-11-22 8e05038
added newer examples to readme file in the examples/ directory
Still need to update the new example headers to correctly point to some documentation of the pysindy doc website
