+++ examples/README.rst
@@ -38,14 +38,38 @@ Explore the differentiation methods available in PySINDy on pure differentiation
 ------------------------------------------------------------------------------------------------------------------------
 See a demonstration of PySINDy objects designed to conform to the `Deeptime <https://deeptime-ml.github.io/latest/index.html>`_ API.
 
-`Plasma physics <https://pysindy.readthedocs.io/en/latest/examples/7_plasma_example.htm>`_
+`Plasma physics <https://pysindy.readthedocs.io/en/latest/examples/7_plasma_example.html>`_
 ----------------------------------------------------------------------------------------------
-See the ``ConstrainedSR3`` optimizer used to build a constrained model for the temporal POD modes of a plasma simulation.
+Use the ``ConstrainedSR3`` optimizer to build a constrained model for the temporal POD modes of a plasma simulation.
 
 
-`Plasma physics <https://pysindy.readthedocs.io/en/latest/examples/8_trapping_sindy_paper_examples.htm>`_
+`Trapping SINDy <https://pysindy.readthedocs.io/en/latest/examples/8_trapping_sindy_paper_examples.html>`_
 ----------------------------------------------------------------------------------------------
-This notebook applies the ``TrappingSR3`` optimizer, proposed in `this paper <https://arxiv.org/abs/2105.01843>`_, to various canonical fluid systems.
+This notebook applies the ``TrappingSR3`` optimizer to various canonical fluid systems., proposed in this paper: Kaptanoglu, Alan A., et al. "Promoting global stability in data-driven models of quadratic nonlinear dynamics." Physical Review Fluids 6.9 (2021): 094401. A preprint is found here `<https://arxiv.org/abs/2105.01843>`_.
+
+`SINDyPI <https://pysindy.readthedocs.io/en/latest/examples/9_sindypi_with_sympy.html>`_
+----------------------------------------------------------------------------------------------
+This notebook applies the ``SINDyPI`` optimizer to a simple implicit ODE and was originally proposed in this paper: Kaheman, Kadierdan, J. Nathan Kutz, and Steven L. Brunton. "SINDy-PI: a robust algorithm for parallel implicit sparse identification of nonlinear dynamics." Proceedings of the Royal Society A 476.2242 (2020): 20200279. 
+
+`PDEFIND <https://pysindy.readthedocs.io/en/latest/examples/10_PDEFIND_examples.html>`_
+----------------------------------------------------------------------------------------------
+This notebook applies the PDEFIND algorithm (SINDy for PDE identification) to a number of PDEs, and was originally proposed in this paper: Rudy, Samuel H., et al. "Data-driven discovery of partial differential equations." Science Advances 3.4 (2017): e1602614.
+
+`Greedy Algorithms <https://pysindy.readthedocs.io/en/latest/examples/11_SSR_FROLS_examples.html>`_
+----------------------------------------------------------------------------------------------
+This notebook uses the step-wise sparse regression (SSR) and forward-regression orthogonal least-squares (FROLS) algorithms, which are greedy algorithms that iteratively truncate (or add) one nonzero coefficient at each algorithm iteration. 
+
+`Weak formulation SINDy <https://pysindy.readthedocs.io/en/latest/examples/12_weakform_SINDy_examples.html>`_
+----------------------------------------------------------------------------------------------
+This notebook uses SINDy to identify the weak-formulation of a system of ODEs or PDEs, adding significant robustness against noise in the data.
+
+`Model ensembles <https://pysindy.readthedocs.io/en/latest/examples/13_ensembling.html>`_
+----------------------------------------------------------------------------------------------
[…]
