### 2021-11-23 987eb62
Added a few more tests for increasing code coverage. Updated notebooks 10, 11, 12 and theyre looking a lot better. Trapping notebook still needs work but im saving this commit so I can switch to my work computer, which runs the notebook muchhhhh faster.
