+++ test/feature_library/test_feature_library.py
@@ -150,20 +150,42 @@ def test_pde_library_bad_parameters(params):
 @pytest.mark.parametrize(
     "params",
     [
-        dict(tensor_array=[[0, 0]]),
-        dict(tensor_array=[[0, 1]]),
-        dict(tensor_array=[[1, -1]]),
-        dict(tensor_array=[[2, 1]]),
-        dict(tensor_array=[1, 1]),
-        dict(tensor_array=[[1, 1, 1]]),
-        dict(inputs_per_library=np.array([[0, 1]])),
-        dict(inputs_per_library=np.array([[0, 1, 2], [0, 1, 2], [0, 1, 2]])),
-        dict(inputs_per_library=np.array([[0, 1, 2], [0, 1, -1]])),
+        dict(libraries=[]),
+        dict(libraries=[PolynomialLibrary, PolynomialLibrary], tensor_array=[[0, 0]]),
+        dict(libraries=[PolynomialLibrary, PolynomialLibrary], tensor_array=[[0, 1]]),
+        dict(libraries=[PolynomialLibrary, PolynomialLibrary], tensor_array=[[1, -1]]),
+        dict(libraries=[PolynomialLibrary, PolynomialLibrary], tensor_array=[[2, 1]]),
+        dict(libraries=[PolynomialLibrary, PolynomialLibrary], tensor_array=[1, 1]),
+        dict(
+            libraries=[PolynomialLibrary, PolynomialLibrary], tensor_array=[[1, 1, 1]]
+        ),
+        dict(
+            libraries=[PolynomialLibrary, PolynomialLibrary],
+            inputs_per_library=np.array([[0, 1], [0, 100]]),
+        ),
+        dict(
+            libraries=[PolynomialLibrary, PolynomialLibrary],
+            inputs_per_library=np.array([0, 0]),
+        ),
+        dict(
+            libraries=[PolynomialLibrary, PolynomialLibrary],
+            inputs_per_library=np.array([[0, 1]]),
+        ),
+        dict(
+            libraries=[PolynomialLibrary, PolynomialLibrary],
+            inputs_per_library=np.array([[0, 1, 2], [0, 1, 2], [0, 1, 2]]),
+        ),
+        dict(
+            libraries=[PolynomialLibrary, PolynomialLibrary],
+            inputs_per_library=np.array([[0, 1, 2], [0, 1, -1]]),
+        ),
     ],
 )
-def test_generalized_library_bad_parameters(params):
+def test_generalized_library_bad_parameters(data_lorenz, params):
     with pytest.raises(ValueError):
-        GeneralizedLibrary([PolynomialLibrary, PolynomialLibrary], **params)
+        lib = GeneralizedLibrary(**params)
+        x, t = data_lorenz
+        lib.fit(x)
 
 
 @pytest.mark.parametrize(
@@ -196,9 +218,11 @@ def test_sindypi_library_bad_params(params):
     [
         IdentityLibrary(),
         PolynomialLibrary(),
+        PolynomialLibrary(include_bias=False),
         FourierLibrary(),
         IdentityLibrary() + PolynomialLibrary(),
         pytest.lazy_fixture("data_custom_library"),
+        pytest.lazy_fixture("data_custom_library_bias"),
         pytest.lazy_fixture("data_generalized_library"),
         pytest.lazy_fixture("data_ode_library"),
         pytest.lazy_fixture("data_pde_library"),
@@ -216,9 +240,11 @@ def test_fit_transform(data_lorenz, library):
     [
         IdentityLibrary(),
         PolynomialLibrary(),
+        PolynomialLibrary(include_bias=False),
         FourierLibrary(),
         IdentityLibrary() + PolynomialLibrary(),
[…]
