+++ test/conftest.py
@@ -208,6 +208,28 @@ def data_custom_library():
 
 
 @pytest.fixture
+def data_custom_library_bias():
+    library_functions = [
+        lambda x: x,
+        lambda x: x ** 2,
+        lambda x: 0 * x,
+        lambda x, y: x * y,
+    ]
+    function_names = [
+        lambda s: str(s),
+        lambda s: str(s) + "^2",
+        lambda s: "0",
+        lambda s, t: str(s) + " " + str(t),
+    ]
+
+    return CustomLibrary(
+        library_functions=library_functions,
+        function_names=function_names,
+        include_bias=True,
+    )
+
+
+@pytest.fixture
 def data_quadratic_library():
     library_functions = [
         lambda x: x,
