### 2021-11-23 ec5124e
Updated the readme with new arxiv paper and new flowchart
