### 2021-11-24 63f337a
Changed odeint to solve_ivp in all the notebooks. Figured out that differences with old notebooks was because solve_ivp defaults to RK45 solvers. Added solve_ivp with LSODA and small error tolerances as the default solver in pysindy.py. Cleaned up a bunch of these notebooks. Fixed the trapping region estimate in the trapping SINDy notebook.
