### 2021-11-24 60bd741
Made some small changes to fix the new sphinx documentation. Updated some of the documentation taglines, fixed small formatting in the REAME and other files, and fixed documentation errors related to numpy.random functions... which for some reason generate their own documentation if you import them directly?
