### 2021-11-27 b9600e2
Significantly simplified the trapping SINDy optimizer usage. Working through the trapping example notebook to make sure everything is reproduced, but need to switch to my desktop so it is only half-finished. Added an example 15 jupyter notebook that will contain the material I plan to go over in upcoming PySINDy youtube tutorials.
