+++ pysindy/optimizers/trapping_sr3.py
@@ -57,12 +57,12 @@ class TrappingSR3(SR3):
         algorithm default is to ignore the stability term. In this limit,
         this should be approximately equivalent to the ConstrainedSR3 method.
 
-    alpha_m : float, optional (default 5e19)
+    alpha_m : float, optional (default eta * 0.1)
         Determines the step size in the prox-gradient descent over m.
         For convergence, need alpha_m <= eta / ||w^T * PQ^T * PQ * w||.
         Typically 0.01 * eta <= alpha_m <= 0.1 * eta.
 
-    alpha_A : float, optional (default 1.0e20)
+    alpha_A : float, optional (default eta)
         Determines the step size in the prox-gradient descent over A.
         For convergence, need alpha_A <= eta, so typically
         alpha_A = eta is used.
@@ -123,13 +123,6 @@ class TrappingSR3(SR3):
         Initial guess for vector A in the optimization. Otherwise
         A is initialized as A = diag(gamma).
 
-    PL : np.ndarray, shape (n_targets, n_targets, n_targets, n_features)
-        Linear coefficient part of the P matrix in ||Pw - A||^2
-
-    PQ : np.ndarray, shape (n_targets, n_targets,
-                            n_targets, n_targets, n_features)
-        Quadratic coefficient part of the P matrix in ||Pw - A||^2
-
     fit_intercept : boolean, optional (default False)
         Whether to calculate the intercept for this model. If set to false, no
         intercept will be used in calculations.
@@ -175,6 +168,16 @@ class TrappingSR3(SR3):
         A^S is indeed being pulled towards the space of negative
         definite matrices.
 
+    PL_unsym_ : np.ndarray, shape (n_targets, n_targets, n_targets, n_features)
+        Unsymmetrized linear coefficient part of the P matrix in ||Pw - A||^2
+
+    PL_ : np.ndarray, shape (n_targets, n_targets, n_targets, n_features)
+        Linear coefficient part of the P matrix in ||Pw - A||^2
+
+    PQ_ : np.ndarray, shape (n_targets, n_targets,
+                            n_targets, n_targets, n_features)
+        Quadratic coefficient part of the P matrix in ||Pw - A||^2
+
     Examples
     --------
     >>> import numpy as np
@@ -202,9 +205,9 @@ class TrappingSR3(SR3):
         eps_solver=1e-7,
         relax_optim=True,
         inequality_constraints=False,
-        eta=1.0e20,
-        alpha_A=1.0e20,
-        alpha_m=5e19,
+        eta=None,
+        alpha_A=None,
+        alpha_m=None,
         gamma=-0.1,
         tol=1e-5,
         tol_m=1e-5,
@@ -217,8 +220,6 @@ class TrappingSR3(SR3):
         copy_X=True,
         m0=None,
         A0=None,
-        PL=None,
-        PQ=None,
         objective_history=None,
         constraint_lhs=None,
         constraint_rhs=None,
@@ -235,6 +236,21 @@ class TrappingSR3(SR3):
         )
         if thresholder.lower() not in ("l1", "l2", "weighted_l1", "weighted_l2"):
             raise ValueError("Regularizer must be (weighted) L1 or L2")
+        if eta is None:
+            print(
+                "eta was not set, so defaulting to eta = 1e20 "
+                "with alpha_m = 1e-2 * eta, alpha_A = eta. Here eta is so "
+                "large that the stability term in the optimization "
+                "will be ignored."
+            )
+            eta = 1e20
+            alpha_m = 1e18
+            alpha_A = 1e20
+        else:
[…]
