### 2021-11-28 13a7143
Fully updated the trapping SINDy notebook to use much simpler algorithm parameters, but with similar performance. Updated Example 15 slightly but plenty of work to do. During updating, found small edge case bug where the ensembling does not work properly for PDE libraries (weak form on or off) used for ODE systems since it assumed that PDE libraries would have spatial grids defined.
