### 2021-11-29 8cbaa74
Polished up Example 15. Fixed some small errors here and there, including: error in the GeneralizedLibrary that was not allowing for feature_names to be passed to this library properly. Added tests for GeneralizedLibrary with feature_names parameter passed or not passed.
