### 2021-12-02 e59cb30
Use dict.get instead of dict.__getitem__
