### 2021-12-02 d694cf7
Switch print statement to warning
