### 2021-12-02 d6aa154
Made the fixes from Brian in the pull request, i.e. fixed the notebooks to not just ignore all warnings (only particular ones), fixed the SR3 examples in the Example 1 notebook (they were being used with the corrupted outlier data so were giving poor results), changed the random seed in Example 3 so that the logistic system is correctly identified, and added some warnings and syntax fixes to the TrappingSR3 optimizer.
