### 2021-12-03 #3845 @sbenthall
@georgiastuart There is no strict timeline for reviews. It is common for reviews to take a month or two. If you think you can get to this in December or January, I'd say your timing is entirely appropriate. May I add you as a reviewer?
