### 2021-12-04 bbd827b
updated README with suggestions for future contributions to the code
