### 2021-12-05 ec26dba
Fix check for trivial constraints
