### 2021-12-05 c80e50f
Fix flowchart, add youtube playlist, reformat references
