### 2021-12-09 #3845 @akaptano
Merged the branch into master, everything is now included in release 1.5.1, so regenerating the pdf here:
