+++ pysindy/feature_library/pde_library.py
@@ -30,129 +30,9 @@ from .base import BaseFeatureLibrary
 from pysindy.differentiation import FiniteDifference
 
 
+#nearing completion, but we need to debug, test, and validate....
 class PDELibrary(BaseFeatureLibrary):
     """Generate a PDE library with custom functions.
-
-    Parameters
-    ----------
-    library_functions : list of mathematical functions, optional (default None)
-        Functions to include in the library. Each function will be
-        applied to each input variable (but not their derivatives)
-
-    derivative_order : int, optional (default 0)
-        Order of derivative to take on each input variable, max 4
-
-    spatial_grid : np.ndarray, optional (default None)
-        The spatial grid for computing derivatives
-
-    temporal_grid : 1D np.ndarray, optional (default None)
-        The temporal grid for integrating the LHS of the SINDy equation
-        when doing a weak-form version of SINDy.
-
-    function_names : list of functions, optional (default None)
-        List of functions used to generate feature names for each library
-        function. Each name function must take a string input (representing
-        a variable name), and output a string depiction of the respective
-        mathematical function applied to that variable. For example, if the
-        first library function is sine, the name function might return
-        :math:`\\sin(x)` given :math:`x` as input. The function_names list
-        must be the same length as library_functions.
-        If no list of function names is provided, defaults to using
-        :math:`[ f_0(x),f_1(x), f_2(x), \\ldots ]`.
-
-    interaction_only : boolean, optional (default True)
-        Whether to omit self-interaction terms.
-        If True, function evaulations of the form :math:`f(x,x)` and :math:`f(x,y,x)`
-        will be omitted, but those of the form :math:`f(x,y)` and :math:`f(x,y,z)`
-        will be included.
-        If False, all combinations will be included.
-
-    include_bias : boolean, optional (default False)
-        If True (default), then include a bias column, the feature in which
-        all polynomial powers are zero (i.e. a column of ones - acts as an
-        intercept term in a linear model).
-        This is hard to do with just lambda functions, because if the system
-        is not 1D, lambdas will generate duplicates.
-
-    include_interaction : boolean, optional (default True)
-        This is a different than the use for the PolynomialLibrary. If true,
-        it generates all the mixed derivative terms. If false, the library
-        will consist of only pure no-derivative terms and pure derivative
-        terms, with no mixed terms.
-
-    is_uniform : boolean, optional (default True)
-        If True, assume the grid is uniform in all spatial directions, so
-        can use uniform grid spacing for the derivative calculations.
-
-    weak_form : boolean, optional (default False)
-        If True, uses the weak/integral form of SINDy. Note that the default
-        is that the user does not need to specify any extra parameters beyond
-        the normal PDE use (i.e. they still need to specify a spatial_grid),
-        although there are many optional parameters below.
-
-    K : int, optional (default 100)
-        Number of domain centers, corresponding to subdomain squares of length
-        Hx and height Hy. If weak_form is True but K is not
-        specified, defaults to 100.
-
-    Hx : float, optional (default None)
-        Half of the length of the square subdomains. If weak_form is True
-        but Hx is not specified, defaults to Hx = Lx / 20, where
[…]
