+++ pysindy/differentiation/finite_difference.py
@@ -2,7 +2,12 @@ import numpy as np
 
 from .base import BaseDifferentiation
 
-
+#cf. Note that in order to maintain compatibility with sklearn the, 
+#array returned, x_dot, always satisfies np.ndim(x_dot) == 2.
+#
+#I implemented differentiation over arbitrary shapes to significantly improve runtime for PDEs
+#The PDE_Library already handles reshaping the feature library to be 2-dimensional for compatibility with sklearn instead
+#Adding different stencils and higher order derivatives wouldn't be too hard either...
 class FiniteDifference(BaseDifferentiation):
     """Finite difference derivatives.
 
@@ -45,7 +50,7 @@ class FiniteDifference(BaseDifferentiation):
            [ 0.53780339, -0.84443737]])
     """
 
-    def __init__(self, order=2, d=1, is_uniform=False, drop_endpoints=False):
+    def __init__(self, order=2, d=1, axis=0, is_uniform=False, drop_endpoints=False):
         if order <= 0 or not isinstance(order, int):
             raise ValueError("order must be a positive int")
         elif order > 2:
@@ -64,6 +69,7 @@ class FiniteDifference(BaseDifferentiation):
         self.order = order
         self.drop_endpoints = drop_endpoints
         self.is_uniform = is_uniform
+        self.axis=axis
 
     def _differentiate(self, x, t):
         """
@@ -86,21 +92,27 @@ class FiniteDifference(BaseDifferentiation):
             t = t[1] - t[0]
 
         x_dot = np.full_like(x, fill_value=np.nan)
-
+        s0 = [slice(dim) for dim in x.shape]
+        s0[self.axis]=slice(None,-1,None)
+        sm1 = [slice(dim) for dim in x.shape]
+        sm1[self.axis]=slice(-1,None,None)
+        sm2 = [slice(dim) for dim in x.shape]
+        sm2[self.axis]=slice(-2,-1,None)
+        sm3 = [slice(dim) for dim in x.shape]
+        sm3[self.axis]=slice(-3,-2,None)
+            
         # Uniform timestep (assume t contains dt)
         if np.isscalar(t):
-            x_dot[:-1, :] = (x[1:, :] - x[:-1, :]) / t
+            x_dot[tuple(s0)] = (np.roll(x,-1,axis=self.axis)-x)[tuple(s0)] / t
             if not self.drop_endpoints:
-                x_dot[-1, :] = (3 * x[-1, :] / 2 - 2 * x[-2, :] + x[-3, :] / 2) / t
+                x_dot[tuple(sm1)]=(3 * x[tuple(sm1)] / 2 - 2 * x[tuple(sm2)] + x[tuple(sm3)] / 2) / t
 
         # Variable timestep
         else:
-            t_diff = t[1:] - t[:-1]
-            x_dot[:-1, :] = (x[1:, :] - x[:-1, :]) / t_diff[:, None]
+            t_diff = np.diff(t)
+            x_dot[tuple(s0)] = (np.roll(x,-1,axis=self.axis)-x)[tuple(s0)] / t_diff[:, None]
             if not self.drop_endpoints:
-                x_dot[-1, :] = (
-                    3 * x[-1, :] / 2 - 2 * x[-2, :] + x[-3, :] / 2
-                ) / t_diff[-1]
+                x_dot[tuple(sm1)]=(3 * x[tuple(sm1)] / 2 - 2 * x[tuple(sm2)] + x[tuple(sm3)] / 2) / t_diff[-1]
 
         return x_dot
 
@@ -122,137 +134,176 @@ class FiniteDifference(BaseDifferentiation):
             d = self.d
 
         x_dot = np.full_like(x, fill_value=np.nan)
+        
[…]

[…]
