+++ pysindy/feature_library/pde_library.py
@@ -1,5 +1,6 @@
 from itertools import combinations
 from itertools import combinations_with_replacement as combinations_w_r
+from itertools import product as iproduct
 
 import numpy.random
 from numpy import array
@@ -15,6 +16,10 @@ from numpy import shape
 from numpy import transpose
 from numpy import zeros
 from numpy import zeros_like
+from numpy import newaxis
+from numpy import concatenate
+from numpy import sum
+from numpy import product
 from scipy.integrate import trapezoid
 from scipy.interpolate import interp1d
 from scipy.interpolate import RectBivariateSpline
@@ -33,6 +38,88 @@ from pysindy.differentiation import FiniteDifference
 #nearing completion, but we need to debug, test, and validate....
 class PDELibrary(BaseFeatureLibrary):
     """Generate a PDE library with custom functions.
+
+    Parameters
+    ----------
+    library_functions : list of mathematical functions, optional (default None)
+        Functions to include in the library. Each function will be
+        applied to each input variable (but not their derivatives)
+
+    derivative_order : int, optional (default 0)
+        Order of derivative to take on each input variable, max 4
+
+    spatial_grid : np.ndarray, optional (default None)
+        The spatial grid for computing derivatives
+
+    temporal_grid : 1D np.ndarray, optional (default None)
+        The temporal grid for integrating the LHS of the SINDy equation
+        when doing a weak-form version of SINDy.
+
+    function_names : list of functions, optional (default None)
+        List of functions used to generate feature names for each library
+        function. Each name function must take a string input (representing
+        a variable name), and output a string depiction of the respective
+        mathematical function applied to that variable. For example, if the
+        first library function is sine, the name function might return
+        :math:`\\sin(x)` given :math:`x` as input. The function_names list
+        must be the same length as library_functions.
+        If no list of function names is provided, defaults to using
+        :math:`[ f_0(x),f_1(x), f_2(x), \\ldots ]`.
+
+    interaction_only : boolean, optional (default True)
+        Whether to omit self-interaction terms.
+        If True, function evaulations of the form :math:`f(x,x)` and :math:`f(x,y,x)`
+        will be omitted, but those of the form :math:`f(x,y)` and :math:`f(x,y,z)`
+        will be included.
+        If False, all combinations will be included.
+
+    include_bias : boolean, optional (default False)
+        If True (default), then include a bias column, the feature in which
+        all polynomial powers are zero (i.e. a column of ones - acts as an
+        intercept term in a linear model).
+        This is hard to do with just lambda functions, because if the system
+        is not 1D, lambdas will generate duplicates.
+
+    include_interaction : boolean, optional (default True)
+        This is a different than the use for the PolynomialLibrary. If true,
+        it generates all the mixed derivative terms. If false, the library
+        will consist of only pure no-derivative terms and pure derivative
+        terms, with no mixed terms.
+
+    is_uniform : boolean, optional (default True)
+        If True, assume the grid is uniform in all spatial directions, so
+        can use uniform grid spacing for the derivative calculations.
+
+    library_ensemble : boolean, optional (default False)
+        Whether or not to use library bagging (regress on subset of the
+        candidate terms in the library)
+
+    ensemble_indices : integer array, optional (default [0])
+        The indices to use for ensembling the library.
+
+    Attributes
+    ----------
+    functions : list of functions
+        Mathematical library functions to be applied to each input feature.
+
+    function_names : list of functions
+        Functions for generating string representations of each library
+        function.
+
+    n_input_features_ : int
+        The total number of input features.
+        WARNING: This is deprecated in scikit-learn version 1.0 and higher so
[…]
