+++ pysindy/feature_library/base.py
@@ -9,7 +9,7 @@ from sklearn.base import TransformerMixin
 from sklearn.utils import check_array
 from sklearn.utils.validation import check_is_fitted
 
-
+#I vectorized the tensor product combinations to speed up runtime a bit
 class BaseFeatureLibrary(TransformerMixin):
     """
     Base class for feature libraries.
@@ -336,12 +336,8 @@ class TensoredLibrary(BaseFeatureLibrary):
         -------
         lib_full : All combinations of the numerical library terms.
         """
-        lib_full = np.zeros((lib_i.shape[0], lib_i.shape[-1] * lib_j.shape[-1]))
-        q = 0
-        for i in range(lib_i.shape[-1]):
-            for j in range(lib_j.shape[-1]):
-                lib_full[:, q] = lib_i[:, i] * lib_j[:, j]
-                q = q + 1
+        lib_full = np.reshape(lib_i[:,:,np.newaxis]*lib_j[:,np.newaxis,:],(lib_i.shape[0], lib_i.shape[-1] * lib_j.shape[-1]))
+        
         return lib_full
 
     def _name_combinations(self, lib_i, lib_j):
