+++ pysindy/differentiation/finite_difference_old.py
@@ -1,262 +1,0 @@
-import numpy as np
-
-from .base import BaseDifferentiation
-
-
-class FiniteDifference(BaseDifferentiation):
-    """Finite difference derivatives.
-
-    For now only first and second order finite difference methods have been
-    implemented.
-
-    Parameters
-    ----------
-    order: int, optional (default 2)
-        The order of the finite difference method to be used.
-        If 1, first order forward difference will be used.
-        If 2, second order centered difference will be used.
-
-    d : int, 1, 2, 3, 4 optional (default 1)
-        The order of derivative to take (d > 3 inaccurate).
-
-    is_uniform : boolean, optional (default False)
-        Parameter to tell the differentiation that, although a 1D or 2D
-        grid is passed, it is uniform so can use dx instead of the full
-        grid array.
-
-    drop_endpoints: boolean, optional (default False)
-        Whether or not derivatives are computed for endpoints.
-        If False, endpoints will be set to np.nan.
-        Note that which points are endpoints depends on the method
-        being used.
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> from pysindy.differentiation import FiniteDifference
-    >>> t = np.linspace(0,1,5)
-    >>> X = np.vstack((np.sin(t),np.cos(t))).T
-    >>> fd = FiniteDifference()
-    >>> fd._differentiate(X, t)
-    array([[ 1.00114596,  0.00370551],
-           [ 0.95885108, -0.24483488],
-           [ 0.8684696 , -0.47444711],
-           [ 0.72409089, -0.67456051],
-           [ 0.53780339, -0.84443737]])
-    """
-
-    def __init__(self, order=2, d=1, is_uniform=False, drop_endpoints=False):
-        if order <= 0 or not isinstance(order, int):
-            raise ValueError("order must be a positive int")
-        elif order > 2:
-            raise NotImplementedError
-
-        if d <= 0 or d > 4:
-            raise ValueError("Derivative order must be " " 1, 2, 3, or 4")
-
-        if d > 1 and order != 2:
-            raise ValueError(
-                "For second or third order derivatives, order must equal 2 "
-                "because only centered differences are implemented"
-            )
-
-        self.d = d
-        self.order = order
-        self.drop_endpoints = drop_endpoints
-        self.is_uniform = is_uniform
-
-    def _differentiate(self, x, t):
-        """
-        Apply finite difference method.
-        """
-        if self.order == 1:
-            return self._forward_difference(x, t)
-        else:
-            return self._centered_difference(x, t)
-
-    def _forward_difference(self, x, t=1):
-        """
-        First order forward difference
-        (and 2nd order backward difference for final point).
-
-        Note that in order to maintain compatibility with sklearn the,
-        array returned, x_dot, always satisfies np.ndim(x_dot) == 2.
-        """
-        if self.is_uniform and not np.isscalar(t):
-            t = t[1] - t[0]
-
-        x_dot = np.full_like(x, fill_value=np.nan)
-
-        # Uniform timestep (assume t contains dt)
-        if np.isscalar(t):
-            x_dot[:-1, :] = (x[1:, :] - x[:-1, :]) / t
-            if not self.drop_endpoints:
-                x_dot[-1, :] = (3 * x[-1, :] / 2 - 2 * x[-2, :] + x[-3, :] / 2) / t
-
-        # Variable timestep
-        else:
-            t_diff = t[1:] - t[:-1]
[…]
