+++ pysindy/feature_library/__init__.py
@@ -5,6 +5,7 @@ from .fourier_library import FourierLibrary
 from .generalized_library import GeneralizedLibrary
 from .identity_library import IdentityLibrary
 from .pde_library import PDELibrary
+from .weak_pde_library import WeakPDELibrary
 from .polynomial_library import PolynomialLibrary
 from .sindy_pi_library import SINDyPILibrary
 
@@ -17,5 +18,6 @@ __all__ = [
     "IdentityLibrary",
     "PolynomialLibrary",
     "PDELibrary",
+    "WeakPDELibrary",
     "SINDyPILibrary",
 ]
