### 2021-12-14 5860799
Made some basic linting changes so far, and slowly adding some unit tests for the new pde functionality.
