+++ pysindy/feature_library/pde_library.py
@@ -2,31 +2,14 @@ from itertools import combinations
 from itertools import combinations_with_replacement as combinations_w_r
 from itertools import product as iproduct
 
-import numpy.random
 from numpy import array
-from numpy import asarray
-from numpy import empty
-from numpy import hstack
-from numpy import linspace
-from numpy import meshgrid
-from numpy import ones
-from numpy import ravel
-from numpy import reshape
-from numpy import shape
-from numpy import transpose
-from numpy import zeros
-from numpy import zeros_like
-from numpy import newaxis
 from numpy import concatenate
-from numpy import sum
+from numpy import empty
+from numpy import newaxis
+from numpy import ones
 from numpy import product
-from scipy.integrate import trapezoid
-from scipy.interpolate import interp1d
-from scipy.interpolate import RectBivariateSpline
-from scipy.interpolate import RegularGridInterpolator
-from scipy.special import comb as n_choose_k
-from scipy.special import hyp2f1
-from scipy.special import poch
+from numpy import reshape
+from numpy import sum
 from sklearn import __version__
 from sklearn.utils import check_array
 from sklearn.utils.validation import check_is_fitted
@@ -49,10 +32,6 @@ class PDELibrary(BaseFeatureLibrary):
 
     spatial_grid : np.ndarray, optional (default None)
         The spatial grid for computing derivatives
-
-    temporal_grid : 1D np.ndarray, optional (default None)
-        The temporal grid for integrating the LHS of the SINDy equation
-        when doing a weak-form version of SINDy.
 
     function_names : list of functions, optional (default None)
         List of functions used to generate feature names for each library
@@ -126,7 +105,6 @@ class PDELibrary(BaseFeatureLibrary):
         library_functions=[],
         derivative_order=0,
         spatial_grid=None,
-        temporal_grid=None,
         interaction_only=True,
         function_names=None,
         include_bias=False,
@@ -140,13 +118,11 @@ class PDELibrary(BaseFeatureLibrary):
         )
         self.functions = library_functions
         self.derivative_order = derivative_order
-        self.temporal_grid = temporal_grid
         self.function_names = function_names
         self.interactions_only = interaction_only
         self.include_bias = include_bias
         self.include_interaction = include_interaction
         self.is_uniform = is_uniform
-        self.weak_form = False
 
         if function_names and (len(library_functions) != len(function_names)):
             raise ValueError(
@@ -155,39 +131,37 @@ class PDELibrary(BaseFeatureLibrary):
             )
         if derivative_order <= 0:
             raise ValueError("The derivative order must be >0")
-            
-        if (spatial_grid is None):
+
+        if spatial_grid is None:
             raise ValueError("Spatial grid required")
-            
-        #list of derivatives
-        indices=()
-        if len(spatial_grid.shape)==1:
-            spatial_grid=reshape(spatial_grid,(len(spatial_grid),1))        
+
+        # list of derivatives
+        indices = ()
+        if len(spatial_grid.shape) == 1:
+            spatial_grid = reshape(spatial_grid, (len(spatial_grid), 1))
         dims = spatial_grid.shape[:-1]
 
-        for i in range(0,len(dims)):
-            indices=indices+(range(derivative_order+1),)
+        for i in range(0, len(dims)):
+            indices = indices + (range(derivative_order + 1),)
 
-        multiindices=[]
+        multiindices = []
         for ind in iproduct(*indices):
-            current=array(ind)
-            if(sum(ind)>0 and sum(ind)<=derivative_order):
+            current = array(ind)
+            if sum(ind) > 0 and sum(ind) <= derivative_order:
                 multiindices.append(current)
-        multiindices=array(multiindices)
[…]
