+++ pysindy/differentiation/finite_difference.py
@@ -2,12 +2,13 @@ import numpy as np
 
 from .base import BaseDifferentiation
 
-#cf. Note that in order to maintain compatibility with sklearn the, 
-#array returned, x_dot, always satisfies np.ndim(x_dot) == 2.
-#
-#I implemented differentiation over arbitrary shapes to significantly improve runtime for PDEs
-#The PDE_Library already handles reshaping the feature library to be 2-dimensional for compatibility with sklearn instead
-#Adding different stencils and higher order derivatives wouldn't be too hard either...
+# Z.N. implemented differentiation over arbitrary shapes to significantly
+# improve runtime for PDEs.
+# The PDE_Library already handles reshaping the feature library to be
+# 2-dimensional for compatibility with sklearn instead.
+# Adding different stencils and higher order derivatives would be easy!
+
+
 class FiniteDifference(BaseDifferentiation):
     """Finite difference derivatives.
 
@@ -23,7 +24,7 @@ class FiniteDifference(BaseDifferentiation):
 
     d : int, 1, 2, 3, 4 optional (default 1)
         The order of derivative to take (d > 3 inaccurate).
-        
+
     axis: int, optional (default 0)
         The axis to differentiate along
 
@@ -72,7 +73,7 @@ class FiniteDifference(BaseDifferentiation):
         self.order = order
         self.drop_endpoints = drop_endpoints
         self.is_uniform = is_uniform
-        self.axis=axis
+        self.axis = axis
 
     def _differentiate(self, x, t):
         """
@@ -96,26 +97,32 @@ class FiniteDifference(BaseDifferentiation):
 
         x_dot = np.full_like(x, fill_value=np.nan)
         s0 = [slice(dim) for dim in x.shape]
-        s0[self.axis]=slice(None,-1,None)
+        s0[self.axis] = slice(None, -1, None)
         sm1 = [slice(dim) for dim in x.shape]
-        sm1[self.axis]=slice(-1,None,None)
+        sm1[self.axis] = slice(-1, None, None)
         sm2 = [slice(dim) for dim in x.shape]
-        sm2[self.axis]=slice(-2,-1,None)
+        sm2[self.axis] = slice(-2, -1, None)
         sm3 = [slice(dim) for dim in x.shape]
-        sm3[self.axis]=slice(-3,-2,None)
-            
+        sm3[self.axis] = slice(-3, -2, None)
+
         # Uniform timestep (assume t contains dt)
         if np.isscalar(t):
-            x_dot[tuple(s0)] = (np.roll(x,-1,axis=self.axis)-x)[tuple(s0)] / t
+            x_dot[tuple(s0)] = (np.roll(x, -1, axis=self.axis) - x)[tuple(s0)] / t
             if not self.drop_endpoints:
-                x_dot[tuple(sm1)]=(3 * x[tuple(sm1)] / 2 - 2 * x[tuple(sm2)] + x[tuple(sm3)] / 2) / t
+                x_dot[tuple(sm1)] = (
+                    3 * x[tuple(sm1)] / 2 - 2 * x[tuple(sm2)] + x[tuple(sm3)] / 2
+                ) / t
 
         # Variable timestep
         else:
             t_diff = np.diff(t)
-            x_dot[tuple(s0)] = (np.roll(x,-1,axis=self.axis)-x)[tuple(s0)] / t_diff[:, None]
+            x_dot[tuple(s0)] = (np.roll(x, -1, axis=self.axis) - x)[tuple(s0)] / t_diff[
+                :, None
+            ]
[…]

[…]
