### 2021-12-18 8718295
Update spectral_derivative.py
shortened line
