+++ pysindy/differentiation/spectral_derivative.py
@@ -4,7 +4,8 @@ from .base import BaseDifferentiation
 
 class SpectralDerivative(BaseDifferentiation):
     """Spectral derivatives. Assumes uniform grid, and utilizes FFT to approximate a derivative.
-    Works well for derivatives in periodic dimensions. Equivalent to a maximal-order finite difference, but runs in O(NlogN).
+    Works well for derivatives in periodic dimensions. 
+    Equivalent to a maximal-order finite difference, but runs in O(NlogN).
 
     Parameters
     ----------
