### 2021-12-18 20ea1a4
Fixed some issue with the new weak form PDE library. Got rid of the weak_form parameter since there is no reason to use this library unless you are using the weak form of it.
