+++ pysindy/feature_library/__init__.py
@@ -5,9 +5,9 @@ from .fourier_library import FourierLibrary
 from .generalized_library import GeneralizedLibrary
 from .identity_library import IdentityLibrary
 from .pde_library import PDELibrary
-from .weak_pde_library import WeakPDELibrary
 from .polynomial_library import PolynomialLibrary
 from .sindy_pi_library import SINDyPILibrary
+from .weak_pde_library import WeakPDELibrary
 
 __all__ = [
     "ConcatLibrary",
