### 2021-12-20 a6cd126
Added some unit tests, updated the notebooks to use the differentiation with axis slicing, added some examples of the slicing, and updated weak form to use the new functionality. So far have a dramatic reduction in the amount of code, but still need to get the mixed derivative (integral) terms working.
