+++ pysindy/feature_library/weak_pde_library.py
@@ -1,24 +1,21 @@
 from itertools import combinations
 from itertools import combinations_with_replacement as combinations_w_r
+from itertools import product as iproduct
 
 import numpy.random
 from numpy import array
-from numpy import asarray
+from numpy import concatenate
 from numpy import empty
-from numpy import hstack
 from numpy import linspace
 from numpy import meshgrid
-from numpy import ravel
+from numpy import newaxis
+from numpy import prod
 from numpy import reshape
-from numpy import shape
-from numpy import transpose
+from numpy import sum as np_sum
+from numpy import take
 from numpy import zeros
-from numpy import zeros_like
 from scipy.integrate import trapezoid
-from scipy.interpolate import interp1d
-from scipy.interpolate import RectBivariateSpline
 from scipy.interpolate import RegularGridInterpolator
-from scipy.special import comb as n_choose_k
 from scipy.special import hyp2f1
 from scipy.special import poch
 from sklearn import __version__
@@ -26,7 +23,6 @@ from sklearn.utils import check_array
 from sklearn.utils.validation import check_is_fitted
 
 from .base import BaseFeatureLibrary
-from pysindy.differentiation import FiniteDifference
 
 
 class WeakPDELibrary(BaseFeatureLibrary):
@@ -41,12 +37,11 @@ class WeakPDELibrary(BaseFeatureLibrary):
     derivative_order : int, optional (default 0)
         Order of derivative to take on each input variable, max 4
 
-    spatial_grid : np.ndarray, optional (default None)
-        The spatial grid for computing derivatives
-
-    temporal_grid : 1D np.ndarray (default None)
-        The temporal grid for integrating the LHS of the SINDy equation
-        when doing a weak-form version of SINDy.
+    spatiotemporal_grid : np.ndarray, optional (default None)
+        The spatial grid for computing derivatives. If using weak form
+        terms with derivatives, must be specified with
+        at least one dimension corresponding to a temporal grid, so that
+        integration by parts can be done in the weak formulation.
 
     function_names : list of functions, optional (default None)
         List of functions used to generate feature names for each library
@@ -88,20 +83,11 @@ class WeakPDELibrary(BaseFeatureLibrary):
         Hx and height Hy. If K is not
         specified, defaults to 100.
 
-    Hx : float, optional (default None)
-        Half of the length of the square subdomains. If
-        Hx is not specified, defaults to Hx = Lx / 20, where
-        Lx is the length of the full domain.
-
-    Hy : float, optional (default None)
-        Half of the height of the square subdomains. If
-        Hy is not specified, defaults to Hy = Ly / 20, where
-        Ly is the height of the full domain.
-
-    Ht : float, optional (default None)
-        Half of the temporal length of the square subdomains.
-        If Ht is not specified, defaults to
-        Ht = M / 20, where M is the number of time points.
+    H_xt : float, optional (default None)
+        Half of the length of the square subdomains in each spatiotemporal
+        direction. If H_xt is not specified, defaults to H_xt = L_xt / 20,
+        where L_xt is the length of the full domain in each spatiotemporal
+        direction.
 
     p : int, optional (default 4)
         Positive integer to define the polynomial degree of the spatial weights
@@ -152,8 +138,7 @@ class WeakPDELibrary(BaseFeatureLibrary):
         self,
         library_functions=[],
         derivative_order=0,
-        spatial_grid=None,
-        temporal_grid=None,
+        spatiotemporal_grid=None,
[…]
