### 2021-12-22 737b1d4
Weak form speedup almost working... quite a bit faster than before but more speedup is available and the mixed terms are still not quite right. Also have a small bug in the finite differencing for d=3 with an axis. Updated the notebooks with the new code.
