+++ pysindy/feature_library/weak_pde_library.py
@@ -612,6 +612,8 @@ class WeakPDELibrary(BaseFeatureLibrary):
                     func_temp = trapezoid(func_temp, x=self.xtgrid_k[k, :, i], axis=0)
                 library_integrals[k, :] = func_temp
 
+            # for term like w * u^2 * u_xx -> (w * u^2)_x * u_x
+            # Need to first interpolate everything onto the subdomains
             # Mixed derivative/non-derivative terms
             if self.include_interaction:
                 # mixed integral terms
@@ -637,143 +639,114 @@ class WeakPDELibrary(BaseFeatureLibrary):
                 nonweak_terms = reshape(
                     nonweak_functions, concatenate([self.grid_dims, [n_library_terms]])
                 )
-                integral_weight_derivs_mixed = empty(
-                    (
-                        self.K,
-                        *(self.grid_ndim * [self.num_pts_per_domain]),
-                        self.num_derivatives,
-                    )
-                )
-                deriv_slices = [slice(None)] * integral_weight_derivs_mixed.ndim
-                derivs_mixed_total = empty(
-                    (
-                        concatenate(
-                            [self.grid_dims, [n_library_terms], [self.num_derivatives]]
-                        )
-                    )
-                )
-                derivs_pure_total = empty(
-                    (
-                        concatenate(
-                            [self.grid_dims, [n_features], [self.num_derivatives]]
-                        )
-                    )
-                )
-                # Note excluding temporal derivatives in the feature library
-                for deriv_ind, multiindex in enumerate(self.multiindices):
-                    derivs_mixed = reshape(
-                        nonweak_functions,
-                        concatenate([self.grid_dims, [n_library_terms]]),
-                    )
-                    derivs_pure = reshape(
-                        x, concatenate([self.grid_dims, [n_features]])
-                    )
-                    for axis in range(self.grid_ndim - 1):
-                        s = [0 for dim in self.spatiotemporal_grid.shape]
-                        s[axis] = slice(self.grid_dims[axis])
-                        s[-1] = axis
-                        d_mixed = floor(multiindex[axis] / 2.0)
-                        if d_mixed > 0:
-                            derivs_mixed = FiniteDifference(
-                                d=d_mixed, axis=axis, is_uniform=self.is_uniform
-                            )._differentiate(
-                                derivs_mixed, self.spatiotemporal_grid[tuple(s)]
-                            )
-                        d_pure = multiindex[axis] - floor(multiindex[axis] / 2.0)
-                        if d_pure > 0:
-                            derivs_pure = FiniteDifference(
-                                d=d_pure, axis=axis, is_uniform=self.is_uniform
-                            )._differentiate(
-                                derivs_pure, self.spatiotemporal_grid[tuple(s)]
-                            )
-                    deriv_slices[-1] = deriv_ind
-                    integral_weight_derivs_mixed[deriv_slices] = self._smooth_ppoly(
-                        d_mixed
-                    ) * (-1) ** (np_sum(d_mixed) % 2)
-                    derivs_mixed_total[deriv_slices] = derivs_mixed
-                    derivs_pure_total[deriv_slices] = derivs_pure
-
-                lib_func_interp = []
-                func_mixed_interp = []
+                nonweak_id = reshape(x, concatenate([self.grid_dims, [n_features]]))
+                func_library_interp = []
                 func_pure_interp = []
                 for j in range(n_library_terms):
-                    lib_func_interp.append(
+                    func_library_interp.append(
                         RegularGridInterpolator(
-                            tuple(self.grid_pts), take(nonweak_terms, j, axis=-1)
+                            tuple(self.grid_pts),
+                            take(nonweak_terms, j, axis=-1),
                         )
                     )
[…]
