### 2021-12-22 cd76ce4
fixed variable timestep d=3 finite difference scale factor
