### 2021-12-22 2e986c5
Added a few more unit tests of the axis differentiation. Cleaned up the weak form code a bit.
