### 2021-12-22 5526bd5
Updated the example notebooks. Everything seems to be working as far as I can tell... but not totally sure about if the weak formulation implementation is fully correct + could be faster.
