### 2021-12-23 #3994 @akaptano
Hi @henrykironde and @tuelwer, thanks again for being willing to review this work! Just letting you know we are working on a the pde_optimization branch (now open as a pull request on the repo) that speeds up the PDE and weak form PDE functionality in the code by a factor of 10 or more by replacing for loops with a lot of numpy vectorization tricks. Feel free to check that out if you're interested or annoyed how slow the PDE-related code is on the main branch. This speedup should facilitate some really advanced SINDy for PDEs in the future.

Happy holidays!
Alan
