### 2021-12-23 dfe1121
added mixed_integral sign correction and casted to tuples for future warnings
