+++ pysindy/feature_library/weak_pde_library.py
@@ -84,7 +84,7 @@ class WeakPDELibrary(BaseFeatureLibrary):
 
     K : int, optional (default 100)
         Number of domain centers, corresponding to subdomain squares of length
-        Hx and height Hy. If K is not
+        Hxt. If K is not
         specified, defaults to 100.
 
     H_xt : array of floats, optional (default None)
@@ -257,9 +257,9 @@ class WeakPDELibrary(BaseFeatureLibrary):
             inds = [slice(None)] * (self.grid_ndim + 1)
             for j in range(self.grid_ndim):
                 inds[j] = 0
-            x1[i] = self.spatiotemporal_grid[inds][i]
+            x1[i] = self.spatiotemporal_grid[tuple(inds)][i]
             inds[i] = -1
-            x2[i] = self.spatiotemporal_grid[inds][i]
+            x2[i] = self.spatiotemporal_grid[tuple(inds)][i]
         return x1, x2
 
     def _set_up_grids(self):
@@ -272,13 +272,11 @@ class WeakPDELibrary(BaseFeatureLibrary):
         xtgrid_k = zeros((self.K, self.num_pts_per_domain, self.grid_ndim))
         xt1_k = domain_centers - self.H_xt
         xt2_k = domain_centers + self.H_xt
-        # mini_grids = []
         for k in range(self.K):
             for i in range(self.grid_ndim):
                 xtgrid_k[k, :, i] = linspace(
                     xt1_k[k, i], xt2_k[k, i], self.num_pts_per_domain
                 )
-                # mini_grids.append(xtgrid_k[k, :, i])
         XT = zeros(
             (self.K, *(self.grid_ndim * [self.num_pts_per_domain]), self.grid_ndim)
         )
@@ -286,7 +284,7 @@ class WeakPDELibrary(BaseFeatureLibrary):
         for k in range(self.K):
             inds_XT[0] = k
             grid_k = array([xtgrid_k[k, :, i] for i in range(self.grid_ndim)])
-            XT[inds_XT] = transpose(
+            XT[tuple(inds_XT)] = transpose(
                 array(meshgrid(*grid_k, indexing="ij")),
                 [*range(1, self.grid_ndim + 1), 0],
             )
@@ -301,7 +299,7 @@ class WeakPDELibrary(BaseFeatureLibrary):
                     if j != i:
                         inds[j] = 0
             inds[-1] = i
-            grid_pts.append(self.spatiotemporal_grid[inds])
+            grid_pts.append(self.spatiotemporal_grid[tuple(inds)])
         self.grid_pts = grid_pts
 
         self.domain_centers_expanded = zeros(
@@ -312,7 +310,7 @@ class WeakPDELibrary(BaseFeatureLibrary):
             for j in range(self.grid_ndim):
                 domain_inds[0] = i
                 domain_inds[-1] = j
-                self.domain_centers_expanded[domain_inds] = domain_centers[i, j]
+                self.domain_centers_expanded[tuple(domain_inds)] = domain_centers[i, j]
 
     @staticmethod
     def _combinations(n_features, n_args, interaction_only):
@@ -369,9 +367,6 @@ class WeakPDELibrary(BaseFeatureLibrary):
             feature_names.append("1")
 
         # Include any non-derivative terms
-        # function_lib_len = 0
-        # if self.functions is not None:
-        #     function_lib_len = len(self.functions)
         for i, f in enumerate(self.functions):
             for c in self._combinations(
                 n_features, f.__code__.co_argcount, self.interaction_only
@@ -566,7 +561,7 @@ class WeakPDELibrary(BaseFeatureLibrary):
             func_inds = [slice(None)] * func_new.ndim
             for k in range(self.K):
                 func_inds[0] = k
[…]
