### 2021-12-24 c60e4b7
debugged the weak mixed derivative terms
