+++ pysindy/differentiation/finite_difference.py
@@ -1,5 +1,3 @@
-import warnings
-
 import numpy as np
 
 from .base import BaseDifferentiation
@@ -15,14 +13,14 @@ class FiniteDifference(BaseDifferentiation):
     ----------
     order: int, optional (default 2)
         The order of the finite difference method to be used.
-        If 1, first order forward difference will be used.
-        If 2, second order centered difference will be used.
+        Currently only centered differences are implemented,
+        and order must be a multiple of 2
 
-    d : int, 1, 2, 3, 4 optional (default 1)
-        The order of derivative to take (d > 3 inaccurate).
+    d : int, optional (default 1)
+        The order of derivative to take.  Must be non-negative.
 
     axis: int, optional (default 0)
-        The axis to differentiate along
+        The axis to differentiate along.
 
     is_uniform : boolean, optional (default False)
         Parameter to tell the differentiation that, although a N-dim
@@ -34,6 +32,12 @@ class FiniteDifference(BaseDifferentiation):
         If False, endpoints will be set to np.nan.
         Note that which points are endpoints depends on the method
         being used.
+
+    periodic: boolean, optional (default False)
+        Whether to use periodic boundary conditions for endpoints.
+        Use forward differences for periodic=False and periodic boundaries
+        with centered differences for periodic=True on the boundaries.
+        No effect if drop_endpoints=True
 
     Examples
     --------
@@ -50,311 +54,231 @@ class FiniteDifference(BaseDifferentiation):
            [ 0.53780339, -0.84443737]])
     """
 
-    def __init__(self, order=2, d=1, axis=0, is_uniform=False, drop_endpoints=False):
-        if order <= 0 or not isinstance(order, int):
-            raise ValueError("order must be a positive int")
-        if d < 1:
-            raise ValueError("differentiation order must be a positive int")
-        elif order > 2:
-            raise NotImplementedError
+    def __init__(
+        self,
+        order=2,
+        d=1,
+        axis=0,
+        is_uniform=False,
+        drop_endpoints=False,
+        periodic=False,
+    ):
 
-        if d >= 4:
-            warnings.warn(
-                "Finite differences of arbitrary order are permitted"
-                " but please note that d >= 4 finite differences"
-                " will dramatically amplify any numerical noise."
-            )
-
-        if d > 1 and order != 2:
-            raise ValueError(
-                "For second or third order derivatives, order must equal 2 "
-                "because only centered differences are implemented"
-            )
+        if order <= 0 or not isinstance(order, int) or order % 2 == 1:
+            raise ValueError("order must be a positive even int")
+        if d < 0:
+            raise ValueError("differentiation order must be a nonnegative int")
 
         self.d = d
         self.order = order
-        self.drop_endpoints = drop_endpoints
         self.is_uniform = is_uniform
         self.axis = axis
+        self.drop_endpoints = drop_endpoints
+        self.periodic = periodic
+        self.n_stencil = int(2 * np.floor((self.d + 1) / 2) - 1 + self.order)
+        self.n_stencil_forward = self.d + self.order
+
+    def _coefficients(self, t):
+        nt = len(t)
+        self.stencil_inds = np.array(
[…]
