+++ test/differentiation/test_differentiation_methods.py
@@ -12,22 +12,6 @@ from pysindy.differentiation.base import BaseDifferentiation
 
 
 # Simplest example: just use an assert statement
-def test_forward_difference_length():
-    x = 2 * np.linspace(1, 100, 100)
-    forward_difference = FiniteDifference(order=1)
-    assert len(forward_difference(x)) == len(x)
-
-    forward_difference_nans = FiniteDifference(order=1, drop_endpoints=True)
-    assert len(forward_difference_nans(x)) == len(x)
-
-
-def test_forward_difference_variable_timestep_length():
-    t = np.linspace(1, 10, 100) ** 2
-    x = 2 * t
-    forward_difference = FiniteDifference(order=1)
-    assert len(forward_difference(x, t) == len(x))
-
-
 def test_centered_difference_length():
     x = 2 * np.linspace(1, 100, 100)
     centered_difference = FiniteDifference(order=2)
@@ -47,20 +31,6 @@ def test_centered_difference_variable_timestep_length():
 # Fixtures: data sets to be re-used in multiple tests
 # data_derivative_1d and data_derivative_2d are defined
 # in ../conftest.py
-
-
-def test_forward_difference_1d(data_derivative_1d):
-    x, x_dot = data_derivative_1d
-    forward_difference = FiniteDifference(order=1)
-    np.testing.assert_allclose(forward_difference(x), x_dot)
-
-
-def test_forward_difference_2d(data_derivative_2d):
-    x, x_dot = data_derivative_2d
-    forward_difference = FiniteDifference(order=1)
-    np.testing.assert_allclose(forward_difference(x), x_dot)
-
-
 def test_centered_difference_1d(data_derivative_1d):
     x, x_dot = data_derivative_1d
     centered_difference = FiniteDifference(order=2)
@@ -75,7 +45,7 @@ def test_centered_difference_2d(data_derivative_2d):
 
 def test_centered_difference_2d_uniform(data_derivative_2d):
     x, x_dot = data_derivative_2d
-    centered_difference = FiniteDifference(order=2)
+    centered_difference = FiniteDifference(order=2, is_uniform=True)
     np.testing.assert_allclose(centered_difference(x), x_dot)
 
 
@@ -102,19 +72,15 @@ def test_centered_difference_xy_yx(data_2dspatial):
     u_xy = np.zeros(u.shape)
     u_yx = np.zeros(u.shape)
     for i in range(100):
-        u_y[i, :, :] = FiniteDifference(order=2, d=1)._centered_difference(
-            u[i, :, :], y_grid
-        )
+        u_y[i, :, :] = FiniteDifference(order=2, d=1)._differentiate(u[i, :, :], y_grid)
     for i in range(50):
-        u_x[:, i, :] = FiniteDifference(order=2, d=1)._centered_difference(
-            u[:, i, :], x_grid
-        )
+        u_x[:, i, :] = FiniteDifference(order=2, d=1)._differentiate(u[:, i, :], x_grid)
     for i in range(100):
-        u_xy[i, :, :] = FiniteDifference(order=2, d=1)._centered_difference(
+        u_xy[i, :, :] = FiniteDifference(order=2, d=1)._differentiate(
             u_x[i, :, :], y_grid
         )
     for i in range(50):
-        u_yx[:, i, :] = FiniteDifference(order=2, d=1)._centered_difference(
+        u_yx[:, i, :] = FiniteDifference(order=2, d=1)._differentiate(
             u_y[:, i, :], x_grid
         )
     np.testing.assert_allclose(u_xy, u_yx)
@@ -138,10 +104,10 @@ def test_centered_difference_hot(data_derivative_2d):
 @pytest.mark.parametrize(
     "data, order",
     [
[…]

[…]
