### 2021-12-27 cd994be
Added an illustration in Example 5 showing that Zachs spectral derivative is much faster than the derivative package. Cleaned up some code here and there. Added some tests of the spectral derivatives which seem to be failing, going to ask Zach about this.
