+++ pysindy/differentiation/spectral_derivative.py
@@ -22,7 +22,7 @@ class SpectralDerivative(BaseDifferentiation):
     >>> import numpy as np
     >>> from pysindy.differentiation import SpectralDerivative
     >>> t = np.arange(0,1,0.1)
-    >>> X = np.vstack((np.sin(t),np.cos(t))).T
+    >>> X = np.vstack((np.sin(t), np.cos(t))).T
     >>> sd = SpectralDerivative()
     >>> sd._differentiate(X, t)
     array([[ 6.28318531e+00,  2.69942771e-16],
@@ -51,13 +51,13 @@ class SpectralDerivative(BaseDifferentiation):
 
         q = np.fft.fft(x, axis=self.axis)
         n = x.shape[self.axis]
-        dims = np.ones(len(x.shape), dtype=int)
+        dims = np.ones(x.ndim, dtype=int)
         dims[self.axis] = n
         freqs = np.zeros(n, dtype=np.complex128)
-        positives = np.arange(int(n / 2 + 1))
+        positives = np.arange(n // 2 + 1)
         negatives = np.setdiff1d(np.arange(n), positives)
-        freqs[: int(n / 2 + 1)] = (positives) * 2 * np.pi / (n * t)
-        freqs[int(n / 2 + 1) :] = (negatives - n) * 2 * np.pi / (n * t)
+        freqs[: n // 2 + 1] = positives * 2 * np.pi / (n * t)
+        freqs[n // 2 + 1 :] = (negatives - n) * 2 * np.pi / (n * t)
 
         if x.dtype is complex:
             return np.fft.ifft(

[…]
