+++ pysindy/differentiation/finite_difference.py
@@ -6,15 +6,12 @@ from .base import BaseDifferentiation
 class FiniteDifference(BaseDifferentiation):
     """Finite difference derivatives.
 
-    For now only first, second, and higher-even-order finite
-    difference methods have been implemented.
-
     Parameters
     ----------
     order: int, optional (default 2)
         The order of the finite difference method to be used.
-        Currently only centered differences are implemented,
-        and order must be a multiple of 2
+        Currently only centered differences are implemented, for even order
+        and left-off-centered differences for odd order
 
     d : int, optional (default 1)
         The order of derivative to take.  Must be non-negative.
@@ -64,7 +61,7 @@ class FiniteDifference(BaseDifferentiation):
         periodic=False,
     ):
 
-        if order <= 0 or not isinstance(order, int) or order % 2 == 1:
+        if order <= 0 or not isinstance(order, int):
             raise ValueError("order must be a positive even int")
         if d < 0:
             raise ValueError("differentiation order must be a nonnegative int")
@@ -76,8 +73,12 @@ class FiniteDifference(BaseDifferentiation):
         self.drop_endpoints = drop_endpoints
         self.periodic = periodic
         self.n_stencil = 2 * ((self.d + 1) // 2) - 1 + self.order
-        # self.n_stencil = int(2 * np.floor((self.d + 1) / 2) - 1 + self.order)
         self.n_stencil_forward = self.d + self.order
+
+        if self.d >= self.n_stencil or self.d > self.n_stencil_forward:
+            raise ValueError(
+                "combination of d and order not implement. use larger order"
+            )
 
     def _coefficients(self, t):
         nt = len(t)
@@ -101,19 +102,34 @@ class FiniteDifference(BaseDifferentiation):
     def _coefficients_boundary_forward(self, t):
         # use the same stencil for each boundary point,
         # but change the evaluation point
-        left = np.arange(self.n_stencil_forward)[:, np.newaxis] * np.ones(
-            (self.n_stencil - 1) // 2, dtype=int
-        )
-        right = (-1 - np.arange(self.n_stencil_forward))[:, np.newaxis] * np.ones(
-            (self.n_stencil - 1) // 2, dtype=int
-        )
+        if self.order % 2 == 0:
+            left = np.arange(self.n_stencil_forward)[:, np.newaxis] * np.ones(
+                (self.n_stencil - 1) // 2, dtype=int
+            )
+            right = (-1 - np.arange(self.n_stencil_forward))[:, np.newaxis] * np.ones(
+                (self.n_stencil - 1) // 2, dtype=int
+            )
+            tinds = np.concatenate(
+                [
+                    np.arange((self.n_stencil - 1) // 2, dtype=int),
+                    np.flip(-1 - np.arange((self.n_stencil - 1) // 2, dtype=int)),
+                ]
+            )
+        else:
+            left = np.arange(self.n_stencil_forward)[:, np.newaxis] * np.ones(
+                (self.n_stencil - 1) // 2, dtype=int
+            )
+            right = (-1 - np.arange(self.n_stencil_forward))[:, np.newaxis] * np.ones(
+                1 + (self.n_stencil - 1) // 2, dtype=int
+            )
+            tinds = np.concatenate(
+                [
+                    np.arange((self.n_stencil - 1) // 2, dtype=int),
+                    np.flip(-1 - np.arange(1 + (self.n_stencil - 1) // 2, dtype=int)),
+                ]
+            )
         self.stencil_inds = np.concatenate([left, right], axis=1)
-        tinds = np.concatenate(
-            [
-                np.arange((self.n_stencil - 1) // 2, dtype=int),
-                np.flip(-1 - np.arange((self.n_stencil - 1) // 2, dtype=int)),
-            ]
-        )
+
[…]
