+++ test/differentiation/test_differentiation_methods.py
@@ -13,6 +13,22 @@ from pysindy.differentiation.base import BaseDifferentiation
 
 
 # Simplest example: just use an assert statement
+def test_forward_difference_length():
+    x = 2 * np.linspace(1, 100, 100)
+    forward_difference = FiniteDifference(order=1)
+    assert len(forward_difference(x)) == len(x)
+
+    forward_difference_nans = FiniteDifference(order=1, drop_endpoints=True)
+    assert len(forward_difference_nans(x)) == len(x)
+
+
+def test_forward_difference_variable_timestep_length():
+    t = np.linspace(1, 10, 100) ** 2
+    x = 2 * t
+    forward_difference = FiniteDifference(order=1)
+    assert len(forward_difference(x, t) == len(x))
+
+
 def test_centered_difference_length():
     x = 2 * np.linspace(1, 100, 100)
     centered_difference = FiniteDifference(order=2)
@@ -32,6 +48,20 @@ def test_centered_difference_variable_timestep_length():
 # Fixtures: data sets to be re-used in multiple tests
 # data_derivative_1d and data_derivative_2d are defined
 # in ../conftest.py
+
+
+def test_forward_difference_1d(data_derivative_1d):
+    x, x_dot = data_derivative_1d
+    forward_difference = FiniteDifference(order=1)
+    np.testing.assert_allclose(forward_difference(x), x_dot)
+
+
+def test_forward_difference_2d(data_derivative_2d):
+    x, x_dot = data_derivative_2d
+    forward_difference = FiniteDifference(order=1)
+    np.testing.assert_allclose(forward_difference(x), x_dot)
+
+
 def test_centered_difference_1d(data_derivative_1d):
     x, x_dot = data_derivative_1d
     centered_difference = FiniteDifference(order=2)
@@ -41,8 +71,7 @@ def test_centered_difference_1d(data_derivative_1d):
 def test_spectral_derivative_1d(data_derivative_periodic_1d):
     t, x, x_dot = data_derivative_periodic_1d
     spectral_derivative = SpectralDerivative()
-    print(spectral_derivative(x, t), x_dot)
-    np.testing.assert_allclose(spectral_derivative(x), x_dot)
+    np.testing.assert_allclose(spectral_derivative(x, t), x_dot, atol=1e-12)
 
 
 def test_centered_difference_2d(data_derivative_2d):
@@ -54,7 +83,7 @@ def test_centered_difference_2d(data_derivative_2d):
 def test_spectral_derivative_2d(data_derivative_periodic_2d):
     t, x, x_dot = data_derivative_periodic_2d
     spectral_derivative = SpectralDerivative()
-    np.testing.assert_allclose(spectral_derivative(x, t), x_dot)
+    np.testing.assert_allclose(spectral_derivative(x, t), x_dot, atol=1e-12)
 
 
 def test_centered_difference_2d_uniform(data_derivative_2d):
@@ -132,7 +161,7 @@ def test_finite_difference(data, order):
 # pytest can also check that methods throw errors when appropriate
 def test_forward_difference_dim():
     x = np.ones((5, 5, 5))
-    forward_difference = FiniteDifference(order=2)
+    forward_difference = FiniteDifference(order=1)
     with pytest.raises(ValueError):
         forward_difference(x)
 
@@ -145,8 +174,6 @@ def test_centered_difference_dim():
 
 
 def test_order_error():
-    with pytest.raises(ValueError):
-        FiniteDifference(order=3)
     with pytest.raises(ValueError):
         FiniteDifference(order=-1)
     with pytest.raises(ValueError):
