### 2021-12-30 4320f5d
Updated most of the code to Zachs suggestion that the reshaping of x_dot and the weak form is done behind-the-scenes. Made some small edits here and there and added a few tests of some of the less used but new differentiation stuff. Looks like spectral derivative might have some minor error. Otherwise, think I just need to update example notebooks 1 and 10 to the new PDE x_dot reshaping and then do some last tests to maximize coverage.
