+++ test/feature_library/test_feature_library.py
@@ -24,7 +24,6 @@ from pysindy.feature_library import WeakPDELibrary
 from pysindy.feature_library.base import BaseFeatureLibrary
 from pysindy.optimizers import SINDyPI
 from pysindy.optimizers import STLSQ
-from pysindy.utils import convert_u_dot_integral
 
 
 def test_form_custom_library():
@@ -549,25 +548,22 @@ def test_generalized_library(data_lorenz):
 
 
 # Helper function for testing PDE libraries
-def pde_library_helper(library, u_flattened, u_dot_flattened, coef_first_dim):
+def pde_library_helper(library, u, coef_first_dim):
     opt = STLSQ(normalize_columns=True, alpha=1e-10, threshold=0)
     model = SINDy(optimizer=opt, feature_library=library)
-    model.fit(u_flattened, x_dot=u_dot_flattened)
+    model.fit(u)
     assert np.any(opt.coef_ != 0.0)
 
     n_features = len(model.get_feature_names())
-    model.fit(
-        u_flattened, x_dot=u_dot_flattened, ensemble=True, n_subset=50, n_models=10
-    )
+    model.fit(u, ensemble=True, n_subset=50, n_models=10)
     assert np.shape(model.coef_list) == (10, coef_first_dim, n_features)
 
-    model.fit(u_flattened, x_dot=u_dot_flattened, library_ensemble=True, n_models=10)
+    model.fit(u, library_ensemble=True, n_models=10)
     assert np.shape(model.coef_list) == (10, coef_first_dim, n_features)
 
 
 def test_1D_pdes(data_1d_random_pde):
-    spatial_grid, u_flattened, u_dot_flattened = data_1d_random_pde
-
+    spatial_grid, u, _ = data_1d_random_pde
     library_functions = [lambda x: x, lambda x: x * x]
     library_function_names = [lambda x: x, lambda x: x + x]
     pde_lib = PDELibrary(
@@ -578,12 +574,11 @@ def test_1D_pdes(data_1d_random_pde):
         include_bias=True,
         is_uniform=True,
     )
-    pde_library_helper(pde_lib, u_flattened, u_dot_flattened, 1)
+    pde_library_helper(pde_lib, u, 1)
 
 
 def test_2D_pdes(data_2d_random_pde):
-    spatial_grid, u_flattened, u_dot_flattened = data_2d_random_pde
-
+    spatial_grid, u, _ = data_2d_random_pde
     library_functions = [lambda x: x, lambda x: x * x]
     library_function_names = [lambda x: x, lambda x: x + x]
     pde_lib = PDELibrary(
@@ -594,12 +589,11 @@ def test_2D_pdes(data_2d_random_pde):
         include_bias=True,
         is_uniform=True,
     )
-    pde_library_helper(pde_lib, u_flattened, u_dot_flattened, 2)
+    pde_library_helper(pde_lib, u, 2)
 
 
 def test_3D_pdes(data_3d_random_pde):
-    spatial_grid, u_flattened, u_dot_flattened = data_3d_random_pde
-
+    spatial_grid, u, _ = data_3d_random_pde
     library_functions = [lambda x: x, lambda x: x * x]
     library_function_names = [lambda x: x, lambda x: x + x]
     pde_lib = PDELibrary(
@@ -610,12 +604,11 @@ def test_3D_pdes(data_3d_random_pde):
         include_bias=True,
         is_uniform=True,
     )
[…]
