### 2021-12-31 1044a74
In the process of trying to reshape things properly for multidimensional > 2 arrays, I basically broke everything. Seemed to have fixed things now, although the PDE results look slightly different than before and I'm not sure why.
