+++ pysindy/feature_library/weak_pde_library.py
@@ -2,21 +2,7 @@ from itertools import combinations
 from itertools import combinations_with_replacement as combinations_w_r
 from itertools import product as iproduct
 
-import numpy.random
-from numpy import array
-from numpy import concatenate
-from numpy import empty
-from numpy import floor
-from numpy import isscalar
-from numpy import linspace
-from numpy import meshgrid
-from numpy import prod
-from numpy import reshape
-from numpy import shape
-from numpy import sum as np_sum
-from numpy import take
-from numpy import transpose
-from numpy import zeros
+import numpy as np
 from scipy.integrate import trapezoid
 from scipy.interpolate import RegularGridInterpolator
 from scipy.special import hyp2f1
@@ -191,8 +177,8 @@ class WeakPDELibrary(BaseFeatureLibrary):
 
         # list of integrals
         indices = ()
-        if len(shape(spatiotemporal_grid)) == 1:
-            spatiotemporal_grid = reshape(
+        if np.array(spatiotemporal_grid).ndim == 1:
+            spatiotemporal_grid = np.reshape(
                 spatiotemporal_grid, (len(spatiotemporal_grid), 1)
             )
         dims = spatiotemporal_grid.shape[:-1]
@@ -205,10 +191,10 @@ class WeakPDELibrary(BaseFeatureLibrary):
 
         multiindices = []
         for ind in iproduct(*indices):
-            current = array(ind)
-            if sum(ind) > 0 and sum(ind) <= derivative_order:
+            current = np.array(ind)
+            if np.sum(ind) > 0 and np.sum(ind) <= derivative_order:
                 multiindices.append(current)
-        multiindices = array(multiindices)
+        multiindices = np.array(multiindices)
         num_derivatives = len(multiindices)
 
         self.num_derivatives = num_derivatives
@@ -222,8 +208,8 @@ class WeakPDELibrary(BaseFeatureLibrary):
         xt1, xt2 = self._get_spatial_endpoints()
         L_xt = xt2 - xt1
         if self.H_xt is not None:
-            if isscalar(self.H_xt):
-                self.H_xt = array(self.grid_ndim * [self.H_xt])
+            if np.isscalar(self.H_xt):
+                self.H_xt = np.array(self.grid_ndim * [self.H_xt])
             if self.grid_ndim != len(self.H_xt):
                 raise ValueError(
                     "The user-defined grid (spatiotemporal_grid) and "
@@ -232,7 +218,7 @@ class WeakPDELibrary(BaseFeatureLibrary):
                     "dimensions. For instance, if spatiotemporal_grid is 4D, "
                     "then H_xt should be a 4D list of the subdomain lengths."
                 )
-            if any(self.H_xt <= zeros(len(self.H_xt))):
+            if any(self.H_xt <= np.zeros(len(self.H_xt))):
                 raise ValueError("Values in H_xt must be a positive float")
             elif any(self.H_xt >= L_xt / 2.0):
                 raise ValueError(
@@ -253,8 +239,8 @@ class WeakPDELibrary(BaseFeatureLibrary):
         self._set_up_grids()
 
     def _get_spatial_endpoints(self):
-        x1 = zeros(self.grid_ndim)
-        x2 = zeros(self.grid_ndim)
+        x1 = np.zeros(self.grid_ndim)
+        x2 = np.zeros(self.grid_ndim)
         for i in range(self.grid_ndim):
             inds = [slice(None)] * (self.grid_ndim + 1)
             for j in range(self.grid_ndim):
@@ -266,28 +252,28 @@ class WeakPDELibrary(BaseFeatureLibrary):
 
     def _set_up_grids(self):
         xt1, xt2 = self._get_spatial_endpoints()
[…]
