### 2022-01-06 2d1567f
Okay think I have finalized everything on the pde_optimization branch before the merge. Added some tests, debugged an issue with PDE/WeakPDEs with the generalizedlibrary and ensembling functionalities, cleaned up some code, reran all the jupyter notebooks.
