+++ pysindy/feature_library/generalized_library.py
@@ -4,6 +4,7 @@ from sklearn.utils import check_array
 from sklearn.utils.validation import check_is_fitted
 
 from .base import BaseFeatureLibrary
+from .weak_pde_library import WeakPDELibrary
 
 
 class GeneralizedLibrary(BaseFeatureLibrary):
@@ -105,6 +106,20 @@ class GeneralizedLibrary(BaseFeatureLibrary):
         )
         if len(libraries) > 0:
             self.libraries_ = libraries
+            weak_libraries = False
+            nonweak_libraries = False
+            for lib in self.libraries_:
+                if isinstance(lib, WeakPDELibrary):
+                    weak_libraries = True
+                else:
+                    nonweak_libraries = True
+            if weak_libraries and nonweak_libraries:
+                raise ValueError(
+                    "At least one of the libraries is a weak form library, "
+                    "and at least one of the libraries is not, which will "
+                    "result in a nonsensical optimization problem. Please use "
+                    "all weak form libraries or no weak form libraries."
+                )
         else:
             raise ValueError(
                 "Empty or nonsensical library list passed to this library."
