### 2022-01-06 0a530f4
Fixed all the binder links to v1.6 and fixed a tiny bit of notation in the PDE and WeakPDE libraries.
