+++ pysindy/feature_library/weak_pde_library.py
@@ -16,7 +16,8 @@ from pysindy.differentiation import FiniteDifference
 
 
 class WeakPDELibrary(BaseFeatureLibrary):
-    """Generate a PDE library with custom functions.
+    """Generate a weak formulation library with custom functions and,
+       optionally, any spatial derivatives in arbitrary dimensions.
 
     Parameters
     ----------
@@ -25,11 +26,12 @@ class WeakPDELibrary(BaseFeatureLibrary):
         applied to each input variable (but not their derivatives)
 
     derivative_order : int, optional (default 0)
-        Order of derivative to take on each input variable, max 4
+        Order of derivative to take on each input variable,
+        can be arbitrary non-negative integer.
 
-    spatiotemporal_grid : np.ndarray, optional (default None)
-        The spatial grid for computing derivatives. If using weak form
-        terms with derivatives, must be specified with
+    spatiotemporal_grid : np.ndarray (default None)
+        The spatiotemporal grid for computing derivatives.
+        This variable must be specified with
         at least one dimension corresponding to a temporal grid, so that
         integration by parts can be done in the weak formulation.
 
