### 2022-01-06 #3994 @akaptano
We have merged the PDE improvements into the main branch as v1.6. I have updated the paper to include Zach N. as an author (who did most of the PDE improvement work). Thanks!
