### 2022-01-07 dynamicslab/pysindy#147 @henrykironde
Include the code of conduct
You could adopt contributor covenant from https://www.contributor-covenant.org and customize it to your needs in the docs directory.

Ref: openjournals/joss-reviews#3994
