### 2022-01-08 ba56ca8
Created a contributor covenant file in the docs/ directory
Addresses Issue #147
