### 2022-01-08 dynamicslab/pysindy#147 @akaptano
Added this... if you are satisfied with the change, go ahead and close this. Thanks!
