### 2022-01-08 dynamicslab/pysindy#147 @akaptano
I should add this was an excellent suggestion that we should have thought of! Thanks again :)
