### 2022-01-10 a0e6f57
fixed multiple_trajectories=True for PDELibrary and WeakPDELibrary for ensemble=False
