+++ pysindy/feature_library/weak_pde_library.py
@@ -160,6 +160,7 @@ class WeakPDELibrary(BaseFeatureLibrary):
         self.H_xt = H_xt
         self.p = p
         self.periodic = periodic
+        self.num_trajectories = 1
 
         if function_names and (len(library_functions) != len(function_names)):
             raise ValueError(
@@ -471,7 +472,9 @@ class WeakPDELibrary(BaseFeatureLibrary):
 
         x = check_array(x)
 
-        n_samples_original, n_features = x.shape
+        n_samples_original_full, n_features = x.shape
+        n_samples_original = n_samples_original_full // self.num_trajectories
+
         if float(__version__[:3]) >= 1.0:
             if n_features != self.n_features_in_:
                 raise ValueError("x shape does not match training shape")
@@ -481,47 +484,145 @@ class WeakPDELibrary(BaseFeatureLibrary):
 
         if self.spatiotemporal_grid is not None:
             n_samples = self.K
+            n_samples_full = self.K * self.num_trajectories
 
-        xp = np.empty((n_samples, self.n_output_features_), dtype=x.dtype)
-
-        # library function terms
-        n_library_terms = 0
-        for f in self.functions:
-            for c in self._combinations(
-                n_features, f.__code__.co_argcount, self.interaction_only
-            ):
-                n_library_terms += 1
-
-        # get original-size library functions before integrating
-        nonweak_functions = np.empty(
-            (n_samples_original, n_library_terms), dtype=x.dtype
+        xp_full = np.empty(
+            (self.num_trajectories, n_samples, self.n_output_features_), dtype=x.dtype
         )
-        library_idx = 0
-        for f in self.functions:
-            for c in self._combinations(
-                n_features, f.__code__.co_argcount, self.interaction_only
-            ):
-                nonweak_functions[:, library_idx] = np.reshape(
-                    f(*[x[:, j] for j in c]), (n_samples_original)
+        x_full = np.reshape(
+            x, np.concatenate([[self.num_trajectories], self.grid_dims, [n_features]])
+        )
+
+        # Loop over each trajectory
+        for trajectory_ind in range(self.num_trajectories):
+            x = np.reshape(x_full[trajectory_ind], (n_samples_original, n_features))
+            xp = np.empty((n_samples, self.n_output_features_), dtype=x.dtype)
+
+            # library function terms
+            n_library_terms = 0
+            for f in self.functions:
+                for c in self._combinations(
+                    n_features, f.__code__.co_argcount, self.interaction_only
+                ):
+                    n_library_terms += 1
+
+            # get original-size library functions before integrating
+            nonweak_functions = np.empty(
+                (n_samples_original, n_library_terms), dtype=x.dtype
+            )
+            library_idx = 0
+            for f in self.functions:
+                for c in self._combinations(
+                    n_features, f.__code__.co_argcount, self.interaction_only
+                ):
+                    nonweak_functions[:, library_idx] = np.reshape(
+                        f(*[x[:, j] for j in c]), (n_samples_original)
+                    )
+                    library_idx += 1
+
+            library_functions = np.empty((n_samples, n_library_terms), dtype=x.dtype)
+            library_idx = 0
+
+            integral_weights = self._smooth_ppoly(np.zeros(self.grid_ndim))
+            func_final = np.zeros(self.K)
+            for f in self.functions:
+                for c in self._combinations(
+                    n_features, f.__code__.co_argcount, self.interaction_only
+                ):
+                    func = f(*[x[:, j] for j in c])
+                    func = np.reshape(func, self.grid_dims)
+                    func_interp = RegularGridInterpolator(tuple(self.grid_pts), func)
+                    for k in range(self.K):
+                        func_new = func_interp(np.take(self.XT, k, axis=0))
[…]
