### 2022-01-11 dynamicslab/pysindy#149 @akaptano
I'm not sure. I've actually seen this error myself, but it is inconsistent... sometimes when I run it, it passes successfully, and other times not (with the same error about the max allowed time). I'll look into it and thanks for your time!
