### 2022-01-11 #3994 @tuelwer
@sbenthall @akaptano This is my review:

This submission deals with a major update to the PySINDy package, which has been previously published in JOSS [1]. PySINDy provides several implementations for the sparse identification of nonlinear dynamical systems, i.e., PySINDy performs sparse linear regression to estimate the function f describing the dynamical system from a dataset. To do so additional features are defined (by applying a library of basis functions to the data) to further improve the fit.

The authors are very clear about the contributions of their new version 1.6:
- The new version of PySINDY is now able to handle systems of partial differential equations: PySINDy can now estimate systems that involve derivatives in up to three spatial dimensions in addition to the time derivative.
- Version 1.6 also implements trapping SINDy algorithm for modeling fluid and plasma flows.
- PySINDy now includes step-wise sparse regression (SSR) and forward regression orthogonal least squares (FROLS)

Remarks:
- In my opinion these new features justify a new submission, however, an editor should check this, as I am not sure whether JOSS has any policies regarding submission for updated versions of previously published packages.
- I had no problems installing the package on MacOS Monterey Version 12.0.1 and on Debian 4.19.208-1, but I got the following warning during installation of the dependencies: `scikit-learn 1.0.2 does not provide the extra 'alldeps'`
- As far as I can see, there are tests covering the new features (e.g., the new PDE module is tested in `conftest.py`, FROLS and SSR are tested in `test_optimizers.py`). I was able to run the tests on MacOS, but I was required to re-install scikit-learn. (This might be related to the above mentioned problem.) One test failed (see issue https://github.com/dynamicslab/pysindy/issues/149).
- The toolbox is very nicely documented and there are many JuPyter notebooks containing extensive examples. The results were reproducible. However, there are some missing dependencies (e.g., matplotlib in notebook '10_PDEFIND_examples') when running the notebooks on binder.
- In some cases the computation of the solution can take some time. It would be good if the users could track the progress of the computation, but I am aware of the fact that this is probably not as easy to implement because PySINDy relies on other optimization toolboxes.
- I liked the new tutorial videos on YouTube that very clearly explain the functionality of PySINDy.
- The repository is quite large (697.3 MB according to MacOS). I would suggest that the authors spend some time looking into this. (Since rewriting the git history can cause more issues, I do not consider the large repository size a deal-breaker.)
- The paper itself is very well-written and gives a nice introduction to the topic. It is quite long: 1637 words (JOSS guidelines suggest a paper length of 250-1000 words.), but given the complex nature of the discussed problem, I feel that the length is appropriate. The authors also did a good job at giving an overview over existing packages for SINDy.

Minor remarks:
- Missing figure in Section ["Documentation"](https://pysindy.readthedocs.io/en/latest/)
- Documentation could be improved by typesetting the math in math mode, e.g., in : https://pysindy.readthedocs.io/en/latest/api/pysindy.optimizers.html#module-pysindy.optimizers.trapping_sr3
- JuPyter notebooks in examples could be renamed such that they are sorted correctly (very minor remark)

Summary of my review:
Overall, the submission is of high quality and I see a lot of value in this new version of PySINDy. That's why I would recommend to accept this submission into JOSS.

References:
[1] de Silva et al., (2020). PySINDy: A Python package for the sparse identification of nonlinear dynamical systems from data. Journal of Open Source Software, 5(49), 2104, https://doi.org/10.21105/joss.02104
