### 2022-01-11 312f0c2
Made a new branch where the generalized trapping SINDy work will go.
