+++ pysindy/optimizers/trapping_generalized_sr3.py
@@ -1,0 +1,780 @@
+import warnings
+
+import cvxpy as cp
+import numpy as np
+from scipy.linalg import cho_factor
+from scipy.linalg import cho_solve
+from sklearn.exceptions import ConvergenceWarning
+
+from ..utils import reorder_constraints
+from .sr3 import SR3
+
+
+class TrappingGeneralizedSR3(SR3):
+    """
+    Generalized trapping variant of sparse relaxed regularized regression.
+    This optimizer can be used to identify systems with
+    stable (bounded) solutions.
+
+    Attempts to minimize the objective function
+
+    .. math::
+
+        0.5\\|y-Xw\\|^2_2 + \\lambda \\times R(w)
+        + 0.5\\|Pw-A\\|^2_2/\\eta + \\delta_0(Cw-d)
+        + \\delta_{\\Lambda}(A)}
+
+    where :math:`R(w)` is a regularization function.
+    See the following references for more details:
+
+        Kaptanoglu, Alan A., et al. "Promoting global stability in
+        data-driven models of quadratic nonlinear dynamics."
+        arXiv preprint arXiv:2105.01843 (2021).
+
+        Zheng, Peng, et al. "A unified framework for sparse relaxed
+        regularized regression: Sr3." IEEE Access 7 (2018): 1404-1423.
+
+        Champion, Kathleen, et al. "A unified sparse optimization framework
+        to learn parsimonious physics-informed models from data."
+        IEEE Access 8 (2020): 169259-169271.
+
+    Parameters
+    ----------
+    evolve_w : bool, optional (default True)
+        If false, don't update w and just minimize over (m, A)
+
+    threshold : float, optional (default 0.1)
+        Determines the strength of the regularization. When the
+        regularization function R is the L0 norm, the regularization
+        is equivalent to performing hard thresholding, and lambda
+        is chosen to threshold at the value given by this parameter.
+        This is equivalent to choosing lambda = threshold^2 / (2 * nu).
+
+    eta : float, optional (default 1.0e20)
+        Determines the strength of the stability term ||Pw-A||^2 in the
+        optimization. The default value is very large so that the
+        algorithm default is to ignore the stability term. In this limit,
+        this should be approximately equivalent to the ConstrainedSR3 method.
+
+    alpha_m : float, optional (default eta * 0.1)
+        Determines the step size in the prox-gradient descent over m.
+        For convergence, need alpha_m <= eta / ||w^T * PQ^T * PQ * w||.
+        Typically 0.01 * eta <= alpha_m <= 0.1 * eta.
+
+    alpha_A : float, optional (default eta)
+        Determines the step size in the prox-gradient descent over A.
+        For convergence, need alpha_A <= eta, so typically
+        alpha_A = eta is used.
+
+    gamma : float, optional (default 0.1)
+        Determines the negative interval that matrix A is projected onto.
+        For most applications gamma = 0.1 - 1.0 works pretty well.
+
+    tol : float, optional (default 1e-5)
+        Tolerance used for determining convergence of the optimization
+        algorithm over w.
+
+    tol_m : float, optional (default 1e-5)
+        Tolerance used for determining convergence of the optimization
+        algorithm over m.
+
+    thresholder : string, optional (default 'L1')
+        Regularization function to use. For current trapping SINDy,
+        only the L1 and L2 norms are implemented. Note that other convex norms
+        could be straightforwardly implemented, but L0 requires
+        reformulation because of nonconvexity.
+
+    thresholds : np.ndarray, shape (n_targets, n_features), optional \
+            (default None)
[…]
