### 2022-01-11 b947826
Fixed errors with PDE and weak libraries with multiple trajectories and ensembling on. Added some tests to test_pysindy.py
