### 2022-01-12 c9797ae
Added matplotlib to requirements so that binder notebooks run properly.
