### 2022-01-12 c2ec158
Updated ReadMe so that flowchart image is linked to
