### 2022-01-12 dynamicslab/pysindy#149 @akaptano
This appears to be a dumb bug because of an old print statement in the file. Deleted this and now cannot reproduce your error, even by running pytest with the same seed ("pytest test/property_tests/test_optimizers_complexity.py --hypothesis-seed=52255670341713864011347036311399653498"). I also added @settings(max_examples=20, deadline=None) to the test_complexity_parameter function since occasionally I was seeing inconsistent timing.
  I'll push this change in a moment, and go ahead and close this issue if it solves the issue on your computer. Thanks!
