### 2022-01-12 ec6a480
Did all the things from the first JOSS review. Fixed the math notation in the optimizer descriptions. Added a verbose option for all the optimizers, tested it, added unit tests, and added simple example in notebook 1. Standardized a bunch of variables between SR3 variants. Fixed the error in test_complexity.
