+++ pysindy/optimizers/constrained_sr3.py
@@ -18,22 +18,22 @@ class ConstrainedSR3(SR3):
 
     .. math::
 
-        0.5\\|y-Xw\\|^2_2 + \\lambda \\times R(v)
-        + (0.5 / nu)\\|w-v\\|^2_2
+        0.5\\|y-Xw\\|^2_2 + \\lambda R(u)
+        + (0.5 / \\nu)\\|w-u\\|^2_2
 
         \\text{subject to}
 
     .. math::
 
-        Cw = d
+        CW = d
 
-    over v and w where :math:`R(v)` is a regularization function, C is a
+    over U and W, where :math:`R(U)` is a regularization function, C is a
     constraint matrix, and d is a vector of values. See the following
     reference for more details:
 
         Champion, Kathleen, et al. "A unified sparse optimization framework
         to learn parsimonious physics-informed models from data."
-        arXiv preprint arXiv:1906.10612 (2019).
+        IEEE Access 8 (2020): 169259-169271.
 
     Parameters
     ----------
@@ -115,6 +115,15 @@ class ConstrainedSR3(SR3):
     inequality_constraints : bool, optional (default False)
         If True, CVXPY methods are used to solve the problem.
 
+    verbose : bool, optional (default False)
+        If True, prints out the different error terms every
+        max_iter / 10 iterations.
+
+    verbose_cvxpy : bool, optional (default False)
+        Boolean flag which is passed to CVXPY solve function to indicate if
+        output should be verbose or not. Only relevant for optimizers that
+        use the CVXPY package in some capabity.
+
     Attributes
     ----------
     coef_ : array, shape (n_features,) or (n_targets, n_features)
@@ -150,6 +159,8 @@ class ConstrainedSR3(SR3):
         initial_guess=None,
         thresholds=None,
         inequality_constraints=False,
+        verbose=False,
+        verbose_cvxpy=False,
     ):
         super(ConstrainedSR3, self).__init__(
             threshold=threshold,
@@ -164,8 +175,10 @@ class ConstrainedSR3(SR3):
             fit_intercept=fit_intercept,
             copy_X=copy_X,
             normalize_columns=normalize_columns,
+            verbose=verbose,
         )
 
+        self.verbose_cvxpy = verbose_cvxpy
         self.reg = get_regularization(thresholder)
         self.use_constraints = (constraint_lhs is not None) and (
             constraint_rhs is not None
@@ -243,13 +256,18 @@ class ConstrainedSR3(SR3):
 
         # default solver is OSQP here but switches to ECOS for L2
         try:
-            prob.solve(max_iter=self.max_iter, eps_abs=self.tol, eps_rel=self.tol)
+            prob.solve(
+                max_iter=self.max_iter,
+                eps_abs=self.tol,
+                eps_rel=self.tol,
+                verbose=self.verbose_cvxpy,
+            )
         # Annoying error coming from L2 norm switching to use the ECOS
         # solver, which uses "max_iters" instead of "max_iter", and
         # similar semantic changes for the other variables.
         except TypeError:
             try:
-                prob.solve(abstol=self.tol, reltol=self.tol)
+                prob.solve(abstol=self.tol, reltol=self.tol, verbose=self.verbose_cvxpy)
             except cp.error.SolverError:
                 print("Solver failed, setting coefs to zeros")
                 xi.value = np.zeros(coef_sparse.shape[0] * coef_sparse.shape[1])
@@ -276,25 +294,48 @@ class ConstrainedSR3(SR3):
         self.history_.append(coef_sparse.T)
         return coef_sparse
 
-    def _objective(self, x, y, coef_full, coef_sparse, trimming_array=None):
+    def _objective(self, x, y, q, coef_full, coef_sparse, trimming_array=None):
         """Objective function"""
+        if q != 0:
[…]
