### 2022-01-13 4996107
Fixed a typo in the differentiation notebook binder link, fixed some math notation in the trapping and constrained optimizers, removed matplotlib from requirements-dev since it is included in requirements now.
