### 2022-01-17 dynamicslab/pysindy#149 @tuelwer
Works now! Thanks for fixing!
