### 2022-01-18 bac8182
Updating paper.md to include Andy as author + mention the derivatives functionality
