+++ docs/JOSS2/paper.md
@@ -16,6 +16,8 @@ authors:
     affiliation: 3
   - name: Kadierdan Kaheman
     affiliation: 3
+  - name: Andy Goldschmidt
+    affiliation: 1
   - name: Jared Callaham
     affiliation: 3
   - name: Charles B. Delahunt
@@ -72,8 +74,9 @@ q_1(t_m) & q_2(t_m) & \cdots & q_n(t_m)
 \end{bmatrix}
 \label{Eq:DataMatrix}.
 \end{eqnarray}
-A matrix of derivatives in time, $\mathbf Q_t$, is defined similarly and can be numerically computed from $\mathbf{Q}$. 
-In this case, Eq. \eqref{eq:sindy_expansion} becomes $\mathbf Q_t \approx \mathbf{\Theta}(\mathbf{Q})\mathbf{\Xi}$ and the goal of the SINDy sparse regression problem is to choose a sparse set of coefficients $\mathbf{\Xi}$ that accurately fits the measured data in $\mathbf Q_t$. We can promote sparsity in the identified coefficients via a sparse regularizer $R(\mathbf{\Xi})$, such as the $l_0$ or $l_1$ norm, and use a sparse regression algorithm such as SR3 [@champion2020unified] to solve the resulting optimization problem,
+A matrix of derivatives in time, $\mathbf Q_t$, is defined similarly and can be numerically computed from $\mathbf{Q}$. PySINDy defaults to second order finite differences for computing derivatives, although a host of more sophisticated methods are now available, including arbitrary order finite differences, Savitzky-Galoy derivatives (i.e. polynomial-filtered derivatives), spectral derivatives with optional filters, arbitary order spline derivatives, and total variational derivatives [@ahnert2007numerical;@chartrand2011numerical;@tibshirani2011solution].
+
+After $\mathbf Q_t$ is obtained, Eq. \eqref{eq:sindy_expansion} becomes $\mathbf Q_t \approx \mathbf{\Theta}(\mathbf{Q})\mathbf{\Xi}$ and the goal of the SINDy sparse regression problem is to choose a sparse set of coefficients $\mathbf{\Xi}$ that accurately fits the measured data in $\mathbf Q_t$. We can promote sparsity in the identified coefficients via a sparse regularizer $R(\mathbf{\Xi})$, such as the $l_0$ or $l_1$ norm, and use a sparse regression algorithm such as SR3 [@champion2020unified] to solve the resulting optimization problem,
 \begin{equation}\label{eq:sindy_regression}
   \text{argmin}_{\boldsymbol\Xi}\|\mathbf Q_t - \boldsymbol\Theta(\mathbf{Q}) \boldsymbol\Xi\|^2 + R(\boldsymbol\Xi).
 \end{equation}
