### 2022-01-18 fa18e73
Updated paper.bib with additional references for the derivative functionality in the paper.md file
