+++ docs/JOSS2/paper.bib
@@ -1,3 +1,34 @@
+@article{ahnert2007numerical,
+  title={Numerical differentiation of experimental data: local versus global methods},
+  author={Ahnert, Karsten and Abel, Markus},
+  journal={Computer Physics Communications},
+  volume={177},
+  number={10},
+  pages={764--774},
+  year={2007},
+  publisher={Elsevier}
+}
+
+@article{chartrand2011numerical,
+  title={Numerical differentiation of noisy, nonsmooth data},
+  author={Chartrand, Rick},
+  journal={International Scholarly Research Notices},
+  volume={2011},
+  year={2011},
+  publisher={Hindawi}
+}
+
+@article{tibshirani2011solution,
+  title={The solution path of the generalized lasso},
+  author={Tibshirani, Ryan J and Taylor, Jonathan},
+  journal={The annals of statistics},
+  volume={39},
+  number={3},
+  pages={1335--1371},
+  year={2011},
+  publisher={Institute of Mathematical Statistics}
+}
+
 @article{delahunt2021toolkit,
   title={A toolkit for data-driven discovery of governing equations in high-noise regimes},
   author={Delahunt, Charles B and Kutz, J Nathan},
