### 2022-01-18 f7cf07a
Updated readme with Andy added as author on new JOSS paper
