### 2022-01-18 #3994 @akaptano
Added Andy Goldschmidt as an author for his work on the derivative package and added couple sentences about the new differentiation methods.
