### 2022-01-19 bb47d5e
Added Urbans ensemble SINDy paper to bib
