+++ docs/JOSS2/paper.bib
@@ -81,6 +81,13 @@
   doi={10.21105/joss.02104}
 }
 
+@article{fasel2021ensemble,
+  title={Ensemble-{SIND}y: {R}obust sparse model discovery in the low-data, high-noise limit, with active learning and control},
+  author={Fasel, Urban and Kutz, J Nathan and Brunton, Bingni W and Brunton, Steven L},
+  journal={arXiv preprint arXiv:2111.10992},
+  year={2021}
+}
+
 @article{fasel2021sindy,
   title={{SIND}y with Control: {A} Tutorial},
   author={Fasel, Urban and Kaiser, Eurika and Kutz, J Nathan and Brunton, Bingni W and Brunton, Steven L},
