### 2022-01-19 fc9bfa0
Added Youtube tutorial link and urban ensembling reference to new JOSS paper
