### 2022-01-19 #3994 @akaptano
Thanks @tuelwer. A (hopefully last) update, made some minor changes to the paper file:
1. Added a link to the YouTube tutorials
2. Updated Figs 1 and 3 to differentiate between PDELibrary and the new WeakPDELibrary, and show the new GeneralizedLibrary class.
3. Added a reference to Urban Fasel's ensemble-SINDy paper.
