### 2022-01-20 5038797
eliminated interpolation in the weak form in favor of analytic weighted sums and vectorized somewhat
