+++ pysindy/feature_library/weak_pde_library.py
@@ -3,10 +3,8 @@ from itertools import combinations_with_replacement as combinations_w_r
 from itertools import product as iproduct
 
 import numpy as np
-from scipy.integrate import trapezoid
-from scipy.interpolate import RegularGridInterpolator
-from scipy.special import hyp2f1
-from scipy.special import poch
+from scipy.special import binom
+from scipy.special import perm
 from sklearn import __version__
 from sklearn.utils import check_array
 from sklearn.utils.validation import check_is_fitted
@@ -138,7 +136,6 @@ class WeakPDELibrary(BaseFeatureLibrary):
         include_interaction=True,
         is_uniform=False,
         K=100,
-        num_pts_per_domain=100,
         H_xt=None,
         p=4,
         library_ensemble=False,
@@ -156,7 +153,6 @@ class WeakPDELibrary(BaseFeatureLibrary):
         self.include_interaction = include_interaction
         self.is_uniform = is_uniform
         self.K = K
-        self.num_pts_per_domain = num_pts_per_domain
         self.H_xt = H_xt
         self.p = p
         self.periodic = periodic
@@ -258,53 +254,62 @@ class WeakPDELibrary(BaseFeatureLibrary):
         self.grid_dims = dims
 
         xt1, xt2 = self._get_spatial_endpoints()
-        domain_centers = np.zeros((self.K, self.grid_ndim))
+        self.domain_centers = np.zeros((self.K, self.grid_ndim))
         for i in range(self.grid_ndim):
-            domain_centers[:, i] = np.random.uniform(
+            self.domain_centers[:, i] = np.random.uniform(
                 xt1[i] + self.H_xt[i], xt2[i] - self.H_xt[i], size=self.K
             )
-        xtgrid_k = np.zeros((self.K, self.num_pts_per_domain, self.grid_ndim))
-        xt1_k = domain_centers - self.H_xt
-        xt2_k = domain_centers + self.H_xt
+
+        self.inds_k = []
         for k in range(self.K):
+            # indices corresponding to the subdomain
+            inds = []
             for i in range(self.grid_ndim):
-                xtgrid_k[k, :, i] = np.linspace(
-                    xt1_k[k, i], xt2_k[k, i], self.num_pts_per_domain
+                s = [0] * (self.grid_ndim + 1)
+                s[i] = slice(None)
+                s[-1] = i
+                newinds = np.intersect1d(
+                    np.where(
+                        self.spatiotemporal_grid[tuple(s)]
+                        >= self.domain_centers[k][i] - self.H_xt[i]
+                    ),
+                    np.where(
+                        self.spatiotemporal_grid[tuple(s)]
+                        <= self.domain_centers[k][i] + self.H_xt[i]
+                    ),
                 )
-        XT = np.zeros(
-            (self.K, *(self.grid_ndim * [self.num_pts_per_domain]), self.grid_ndim)
-        )
-        inds_XT = [slice(None)] * XT.ndim
+                if len(newinds) < 2:
+                    for i in range(self.grid_ndim):
+                        self.domain_centers[k, i] = np.random.uniform(
+                            xt1[i] + self.H_xt[i], xt2[i] - self.H_xt[i], size=1
+                        )
+                    k -= 1
+                    break
+                else:
+                    inds = inds + [newinds]
+            self.inds_k = self.inds_k + [inds]
+
+        self.XT_k = [
+            self.spatiotemporal_grid[np.ix_(*self.inds_k[k])] for k in range(self.K)
+        ]
+
+        self.H_xt_k = np.zeros((self.K, self.grid_ndim))
+        # self.H_xt_k[:] = np.reshape(self.H_xt, (1, self.grid_ndim))
[…]
