+++ pysindy/utils/__init__.py
@@ -1,5 +1,4 @@
 from .base import capped_simplex_projection
-from .base import convert_u_dot_integral
 from .base import drop_nan_rows
 from .base import drop_random_rows
 from .base import equations
@@ -42,6 +41,12 @@ from .odes import pendulum_on_cart
 from .odes import rossler
 from .odes import van_der_pol
 from .odes import yeast
+
+# from .base import convert_u_dot_integral
+# from .base import integrate
+# from .base import integrate2
+# from .base import phi
+# from .base import linear_weights
 
 __all__ = [
     "capped_simplex_projection",
