### 2022-01-20 667524a
Added does for the new references in the new JOSS paper
