+++ docs/JOSS2/paper.bib
@@ -6,7 +6,8 @@
   number={10},
   pages={764--774},
   year={2007},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cpc.2007.03.009}
 }
 
 @article{chartrand2011numerical,
@@ -15,7 +16,8 @@
   journal={International Scholarly Research Notices},
   volume={2011},
   year={2011},
-  publisher={Hindawi}
+  publisher={Hindawi},
+  doi={10.5402/2011/164564}
 }
 
 @article{tibshirani2011solution,
@@ -26,7 +28,8 @@
   number={3},
   pages={1335--1371},
   year={2011},
-  publisher={Institute of Mathematical Statistics}
+  publisher={Institute of Mathematical Statistics},
+  doi={10.1214/11-aos878}
 }
 
 @article{delahunt2021toolkit,
