### 2022-01-20 fe82e95
cleaned up code and added some comments
