+++ pysindy/feature_library/weak_pde_library.py
@@ -260,9 +260,9 @@ class WeakPDELibrary(BaseFeatureLibrary):
                 xt1[i] + self.H_xt[i], xt2[i] - self.H_xt[i], size=self.K
             )
 
+        # Find the indices for space-time points that lie in the domain cells
         self.inds_k = []
         for k in range(self.K):
-            # indices corresponding to the subdomain
             inds = []
             for i in range(self.grid_ndim):
                 s = [0] * (self.grid_ndim + 1)
@@ -278,6 +278,7 @@ class WeakPDELibrary(BaseFeatureLibrary):
                         <= self.domain_centers[k][i] + self.H_xt[i]
                     ),
                 )
+                # If less than two indices along any axis, resample
                 if len(newinds) < 2:
                     for i in range(self.grid_ndim):
                         self.domain_centers[k, i] = np.random.uniform(
@@ -289,12 +290,13 @@ class WeakPDELibrary(BaseFeatureLibrary):
                     inds = inds + [newinds]
             self.inds_k = self.inds_k + [inds]
 
+        # Values of the spatiotemporal grid on the domain cells
         self.XT_k = [
             self.spatiotemporal_grid[np.ix_(*self.inds_k[k])] for k in range(self.K)
         ]
 
+        # Size of the recentered domain cells
         self.H_xt_k = np.zeros((self.K, self.grid_ndim))
-        # self.H_xt_k[:] = np.reshape(self.H_xt, (1, self.grid_ndim))
         for k in range(self.K):
             for axis in range(self.grid_ndim):
                 s = [0] * (self.grid_ndim + 1)
@@ -306,16 +308,124 @@ class WeakPDELibrary(BaseFeatureLibrary):
                 self.domain_centers[k][axis] = (
                     self.XT_k[k][tuple(s)][-1] + self.XT_k[k][tuple(s)][0]
                 ) / 2
+        # Rescaled space-time values for integration weights
         self.xtilde_k = [
             (self.XT_k[k] - self.domain_centers[k]) / self.H_xt_k[k]
             for k in range(self.K)
         ]
+
+        # Integration weights for the library function terms
+        self.weights0 = []
+        for k in range(self.K):
+            self.weights0 = self.weights0 + [
+                self._get_weights(self.xtilde_k[k], np.zeros(self.grid_ndim), self.p)
+                * np.product(self.H_xt_k[k])
+            ]
+
+        # Integration weights for the library integral terms
+        self.weights1 = []
+        for k in range(self.K):
+            weights2 = []
+            for j in range(self.num_derivatives):
+                deriv = np.concatenate([self.multiindices[j], [0]])
+                weights = self._get_weights(
+                    self.xtilde_k[k], deriv, self.p
+                ) * np.product(self.H_xt_k[k] ** (1.0 - deriv))
+                weights2 = weights2 + [weights]
+            self.weights1 = self.weights1 + [weights2]
 
     @staticmethod
     def _combinations(n_features, n_args, interaction_only):
         """Get the combinations of features to be passed to a library function."""
         comb = combinations if interaction_only else combinations_w_r
         return comb(range(n_features), n_args)
+
+    def _phi(self, x, d, p):
+        """One-dimensional polynomial test function (1-x**2)**p,
+        differentiated d times, calculated term-wise in the binomial
+        expansion."""
+        ks = np.arange(self.p + 1)
+        ks = ks[np.where(2 * (self.p - ks) - d >= 0)][:, np.newaxis]
+        return np.sum(
+            binom(self.p, ks)
+            * (-1) ** ks
+            * x[np.newaxis, :] ** (2 * (self.p - ks) - d)
+            * perm(2 * (self.p - ks), d),
+            axis=0,
+        )
+
+    def __phi_int(self, x, d, p):
[…]
