+++ pysindy/feature_library/pde_library.py
@@ -186,7 +186,7 @@ class PDELibrary(BaseFeatureLibrary):
         if float(__version__[:3]) >= 1.0:
             n_features = self.n_features_in_
         else:
-            n_features = self.n_input_features
+            n_features = self.n_input_features_
 
         if input_features is None:
             input_features = ["x%d" % i for i in range(n_features)]
