### 2022-01-21 b20099f
recycled function evaluation
