### 2022-01-25 e25dac2
improved weight calculation and vectorized a bit
