+++ pysindy/feature_library/weak_pde_library.py
@@ -262,7 +262,8 @@ class WeakPDELibrary(BaseFeatureLibrary):
 
         # Find the indices for space-time points that lie in the domain cells
         self.inds_k = []
-        for k in range(self.K):
+        k = 0
+        while k < self.K:
             inds = []
             for i in range(self.grid_ndim):
                 s = [0] * (self.grid_ndim + 1)
@@ -284,11 +285,14 @@ class WeakPDELibrary(BaseFeatureLibrary):
                         self.domain_centers[k, i] = np.random.uniform(
                             xt1[i] + self.H_xt[i], xt2[i] - self.H_xt[i], size=1
                         )
-                    k -= 1
+                    include = False
                     break
                 else:
+                    include = True
                     inds = inds + [newinds]
-            self.inds_k = self.inds_k + [inds]
+            if include:
+                self.inds_k = self.inds_k + [inds]
+                k = k + 1
 
         # Values of the spatiotemporal grid on the domain cells
         self.XT_k = [
@@ -314,24 +318,86 @@ class WeakPDELibrary(BaseFeatureLibrary):
             for k in range(self.K)
         ]
 
-        # Integration weights for the library function terms
-        self.weights0 = []
-        for k in range(self.K):
-            self.weights0 = self.weights0 + [
-                self._get_weights(self.xtilde_k[k], np.zeros(self.grid_ndim), self.p)
-                * np.product(self.H_xt_k[k])
+        self.shapes_k = np.array(
+            [
+                [len(self.inds_k[k][i]) for i in range(self.grid_ndim)]
+                for k in range(self.K)
+            ]
+        )
+        self.grids = []
+        self.lefts = []
+        self.rights = []
+        for i in range(self.grid_ndim):
+            s = [0] * (self.grid_ndim + 1)
+            s[-1] = i
+            s[i] = slice(None)
+            self.grids = self.grids + [
+                np.hstack([self.xtilde_k[k][tuple(s)] for k in range(self.K)])
+            ]
+            self.rights = self.rights + [np.cumsum(self.shapes_k[:, i]) - 1]
+            self.lefts = self.lefts + [
+                np.concatenate([[0], np.cumsum(self.shapes_k[:, i])[:-1]])
             ]
 
-        # Integration weights for the library integral terms
+        self.tweights = []
+        deriv = np.zeros(self.grid_ndim)
+        deriv[-1] = 1
+        for i in range(self.grid_ndim):
+            self.tweights = self.tweights + [
+                self._linear_weights(self.grids[i], deriv[i], self.p)
+            ]
+            self.tweights[i][self.lefts[i]] = self._left_weights(
+                self.grids[i][self.lefts[i]],
+                self.grids[i][self.lefts[i] + 1],
+                deriv[i],
+                self.p,
+            )
+            self.tweights[i][self.rights[i]] = self._right_weights(
+                self.grids[i][self.rights[i] - 1],
+                self.grids[i][self.rights[i]],
+                deriv[i],
+                self.p,
+            )
+
+        self.weights0 = []
+        deriv = np.zeros(self.grid_ndim)
+        for i in range(self.grid_ndim):
+            self.weights0 = self.weights0 + [
+                self._linear_weights(self.grids[i], deriv[i], self.p)
+            ]
+            self.weights0[i][self.lefts[i]] = self._left_weights(
+                self.grids[i][self.lefts[i]],
+                self.grids[i][self.lefts[i] + 1],
+                deriv[i],
+                self.p,
+            )
+            self.weights0[i][self.rights[i]] = self._right_weights(
+                self.grids[i][self.rights[i] - 1],
+                self.grids[i][self.rights[i]],
+                deriv[i],
+                self.p,
+            )
+
[…]
