### 2022-01-25 #3994 @henrykironde
software is well packaged to the standard and satisfies substantial scholarly effort. However, we should confirm the submission as @tuelwer mentioned

> an editor should check this, as I am not sure whether JOSS has any policies regarding submission for updated versions of previously published packages.

@akaptano thank you for submitting this package. The YouTube videos on PySINDy, made everything move smoothly, and I am sure the users find it easy to use the tool.
