### 2022-01-26 eb068bc
corrected tiny typo in new JOSS paper
