### 2022-01-26 c92f03e
Fixed another type in JOSS paper
