### 2022-01-26 0be5400
more vectorization
