### 2022-01-27 #3994 @sbenthall
Ok. I've checked in with the editors-in-chief, and resubmission is allowed if there's substantial scholarly effort and a genuinely new major release, which this is.

@akaptano, can you please:
 - make a tagged release
 - make an archive with a DOI on Zenodo or the like
 - report the version number and DOI here
