### 2022-01-27 #3994 @akaptano
@sbenthall The new release is v1.6.2 and the DOI Zenodo archive is [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5842612.svg)](https://doi.org/10.5281/zenodo.5842612)

DOI: 10.5281/zenodo.5842612
