### 2022-01-28 44fdedb
added small fix to generalized library for the weak case
